[roll] Roll fuchsia [magma] Move C API notification sending to the calling thread.
Writing the the channel is thread-safe, so we don't have to post it to
the connection thread. This should at least 16 us of work per
notification on Nelson.

This patch is a port of I0e7df0a1141ad45cfc55785176c73d26bd5f4a0f to
the MSD C API, so it's easier to backport to release branches.

Original-Bug: 122629
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/813601
Original-Revision: d8a19b79a2351df447e8c132dfc32f3ca98b8b05
GitOrigin-RevId: 68fda5d4e9b113d902bc601d7890ee2074ec267c
Change-Id: I8b7100f1a6e06fc7704b78af997d001c55065829