[roll] Roll fuchsia [netstack][dhcp] Wait for packets to be read
Instead of waiting for the server to send a packet, wait until the DHCP
client has read and started processing it. This eliminates flakiness due
to races between processing of a delivered packet and signalling of
timeouts by forcing the latter to be done only after the former is
guaranteed to have started.

Original-Bug: 122637
Multiply: netstack-dhcp-gotests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/811407
Original-Revision: ab179bda0154c886571d993a833adcbbb72e8e1d
GitOrigin-RevId: c0c1260b163df18e34b74a561b24d592c50677ac
Change-Id: I95fe89ad78a9506a57d22b6fcfd9c0e010fc8192