[roll] Roll fuchsia [benchmarks] Implement dumpFile() using SSH instead of SL4F
Add an implementation of dumpFile() that uses SSH instead of the SL4F
server, and use it for the common cases.

This will be useful for avoiding the dependency on SL4F for the
simpler performance tests that just run a component and don't use
tracing.  For now, those simpler tests still start the SL4F server,
but I will address that in a later CL.

Original-Bug: 122165
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/813447
Original-Revision: 02a9095a06db00360c89790185ecd70f99d273c2
GitOrigin-RevId: f0462db403cdd53591c3cccfad95417eecd59816
Change-Id: Ifc982b289fbf840f90712e9b10baaea66aa525f0