[roll] Roll infra/recipes [kythe][rust] Turn down rust indexing for Fuchsia
The save-analysis feature was retracted from the Rust compiler,
and we can no longer get the source code info we need for
analyses.

We are turning the rust indexing down, with the hope that some day
we could re-enable it.

Original-Bug: 123061
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/814389
Original-Revision: b33f573dcdd62fe0dc48cb33cf1ea8cb420cffb6
GitOrigin-RevId: 3ebba30d9234155ded37873fe03eda5d42c3c2a9
Change-Id: I8dd08431e332c28375984399c484e47f5f552118