[roll] Roll fuchsia [fdf][testing] TestNode use default dispatcher
This makes TestNode compatible with the DispatcherBound type by giving
it a default dispatcher if a dispatcher is not passed in. This is
because we cannot manually pass in a dispatcher pointer through
DispatcherBound as raw pointers are not allowed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/813807
Original-Revision: 5ebaa67e9cd10b720d5ac0730bbd4b5679d3f2cb
GitOrigin-RevId: 78f0e5b0e48355b421d6cee8b9a49e64b7acac91
Change-Id: Ia1098fda2f76ae22f0427bcfa81cfa5ad5e7ca50