[roll] Roll fuchsia [wlan][wlancfg] Fix counter stats all 0's bug
When UpdateExperiment is called, new instances of Cobalt metric logger
and stats are created, but the LazyNodes in Inspect are not updated to
point to the new instance of the stats. In this commit, we only refresh
the Cobalt logger and not the stats so that the bug no longer occurs.

However, this only fits with the current usage pattern where
UpdateExperiment is only called at the beginning. It's possible that in
the future it'll be called long after device bootup, at which point this
implementation is no longer sufficient because non-experiment and
experiment stats are intermingled. See fxbug.dev/122985 for discussion
on handling this long term.

This commit also includes a new unit test, and the Inspect LazyNodes
with the counters are also now kept in StatsLogger, to reduce the chance
of introducing the same bug again in the future.

Original-Bug: 120678
Test:
1. Added unit test
2. Verified bug is fixed on nelson

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/814212
Original-Revision: e3a3f2377d9b7b670095a575433d12083248f58d
GitOrigin-RevId: 3f22180dd5920c4ae824c55893e1d258e33e7d63
Change-Id: I6a41e5c6a2ee6d89b2281dca870935eff01c9c71