[roll] Roll fuchsia [virtio-block] Fix invoking block impl callback twice.
In the corner case of performing a block impl IO of 0 blocks, the block
impl callback is invoked twice; once with an error code and the next
with a success code (ZX_OK).

Fix this while using block::CheckIoRange() to check the validity of the
IO range.

Original-Bug: 43065
Test: fx test virtio-block-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/814454
Original-Revision: 89d8f7f33e837adafef7b440d8996426e2da3083
GitOrigin-RevId: 38ff74bb9573cce95fa1fff23dcca33bef1cac8f
Change-Id: Id5589cf5d5c1aac2ec06a420f4abcdb97420e6df