[roll] Roll fuchsia [memory_sampler] Add a service to collect memory profile samples
This commit exposes a new FIDL interface for a process to report its
memory allocations, as well as a component which serves this FIDL
interface.

The component collects allocation information from its client's
process and, once the client stops, creates a `pprof`-compatible
protobuf profile that it persists to a file.

This is expected to be the first change of a series; the component is
currently not routed nor used anywhere, but would eventually expose a
generic mechanism for arbitrary components to opt into profiling.

Test: fx test memory_sampler
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/801327
Original-Revision: d5ee1bf6fafa9a255322a304d8690e2ef01e7353
GitOrigin-RevId: 9ea17ae43bc5e8811325191f3f04761483ab2404
Change-Id: I6f8353d41687470a469ab9b28d1b49b1f84af844