[roll] Roll fuchsia [botanist] Split Target.TestConfig into BaseTarget
Break apart TestConfig from the Target interface into a smaller
interface that allows for a partial implementation. The resulting
slimmed-down interface is embedded into Target for backwards
compatibility with current uses of Target.

This is a necessary step to support generic targets in an upcoming CL.
Generic targets are not guaranteed to be running Fuchsia, but botanist
still needs to export information about them to the testbed config for
tests to injest.

Rename variables to make the type of target clear. There will future
changes to reduce naming redundancy.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/813948
Original-Revision: 7fba8154165cad0fe3399e88eb8c464a831b2ea7
GitOrigin-RevId: d33ca07bcb3b033e7a0d9b2a0bc324492d878dc7
Change-Id: Ic48f5bd07ee0ca976304f84e25d511092e77c9d1