[roll] Roll fuchsia [fx][pkg] Check if the server is running
We now have a way to run the ffx repository server on arbitrary ports,
as well as a way to check if the repository server is running. This
changes `fx serve-updates`, `fx add-update-source`, and `fx-serve-remote`
to use these tools to allow the port to be configured, and to only error
out if we actually have a port conflict, rather than a potential conflict
in in between a local and remote repository server config.

Original-Bug: 82788
Original-Bug: 119089
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791846
Original-Revision: e95c62f7a64693e6981f7ef7232ea21215b4dfef
GitOrigin-RevId: 94ce740bbe0ec11c3ba73d23f51275bb8107e258
Change-Id: I06ad82cc3600a39d7cda266b8e30f4dedbdf71df