[roll] Roll third_party/ffmpeg [ffmpeg] roll ffmpeg to 346d30b3f6631f2d47fa310f7dfeb78215706fc3
This CL rolls ffmpeg forward to use a current version of chromium's
third_party/ffmpeg repo. This is the first roll using the new roll
scripts and the fuchsia main GN files that specify what features to
include.

For the time being, we still have the 'default' and 'max' profiles
that have existing for some time. 'max' has been the profile used
by default for both prebuilds and use_ffmpeg_prebuilt=false builds.

We want the 'default' profile to be the default profile going
forward, and for 'max' to go away. We will also introduce product-
specific profiles. In order to start using 'default' as the default
profile, the 'default' profile has been updated to be identical to
the 'max' profile, so that transition can occur without disruption.
Both profiles differ in benign ways from the previous 'max' profile.

In addition to the roll, this CL makes the following changes:

1) Fixed an error in the roll instructions.
2) Updated to the new lib version numbers.
3) Made the non-prebuild build use 'default' by default.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/third_party/ffmpeg/+/807596
Original-Revision: 51adab68ac27bbd8d1c3fab89c3ced69e8828564
GitOrigin-RevId: 3504f3876b6e6137910a3c69ed75ba59a961a53b
Change-Id: I0cf3c250eb9d65a80bf552fa3378633f0d5720e2