[roll] Roll fuchsia [kernel][vm] Explicit paged property
Makes is_paged a property stored in the VmObject instead of a virtual
method. The VmObject has unused padding such that this does not
increase the size of the object.

Using a stateful property provides a potential performance improvement
as well as smaller code generation. These benefits are not realized at
the moment as is_paged is not presently on any critical path of a
benchmark.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/795484
Original-Revision: c0e574937ded9bad6a98b90cf224d3ae1da4345f
GitOrigin-RevId: ca7f20522eee5402ce4078073a89315bad1883a1
Change-Id: I7ea9477698883c075cd35035b39424aacbec2448