[roll] Roll fuchsia [web_runner_tests] Use hermetic debug netstack
Hermetic netstack to not be affect and be affected by the rest of
the (non-test) system and its netstack.

Debug netstack to debug a flake in web runner tests where it is unable
to open a socket.

Original-Bug: 122169
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/815575
Original-Revision: b6847cec5d1d20f83cdf88ad43ac7bb69c3e6c75
GitOrigin-RevId: 9c517f3e26fea13d65b2bc69244f2150ee142bb1
Change-Id: I3b3066a1c0a6bddce3f289322ebaa5b243bd479d