[roll] Roll fuchsia [netstack3] Introduce TX device queue
This change introduces a queue that sits between the stack and the
device to hold packets while the device is not ready to transmit
buffers. The queue is implemented fully in core and the bindings
is simply expected to support waking up a TX task which is
responsible for kicking the core to send packets to the device.

Note that this change changes the semantics of
`DeviceLayerEventDispatcher::send_frame` to always accept a buffer
and not a `Serializer`. This is because the queue will hold serialized
buffers for the device. Future changes (see https://fxbug.dev/105927)
will leverage a custom allocator to serialize frames directly into a
device's data VMO, reducing data copies in the TX path.

Later changes will update the implementation of
`DeviceLayerEventDispatcher::send_frame` to not drop frames when TX
buffers are unavailable. See https://fxbug.dev/105921.

Note that we do not yet use the TX device queues. The relevant changes
to bindings will come in a follow-up change.

Original-Bug: 105615, 105921, 105927

test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/814227
Original-Revision: ada91334beae55e7c6c57797747161899089f6e1
GitOrigin-RevId: b2060d90238c7910c2657cd9c67e3a3142b54d7f
Change-Id: Ib217111f1663df96c2d1d9973a11a2fb0dd5e5ab