[roll] Roll fuchsia [build] Change build/cat.sh to atomically write files
This changes the build to atomically write the all_package_manifests.list.
This is primarily useful for the incremental publisher, which watches
this file for changes. Since this file format is line oriented, there
isn't a way to know if the file was truncated, or fully written. This
means that if the incremental publisher manages to read the file before
it's fully written, it will only try to publish a subset of changes. By
atomically writing this file, we should be protected from this state.

Long term though we should change the file format so we can tell if
this file was truncated to catch this error with out-of-tree package
manifest lists.

Original-Fixed: 123151
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/815272
Original-Revision: c3830187eb271b4228598b75e71f2f8c423b04e8
GitOrigin-RevId: 418062c1111b2d7e7eac5aa1641a1f76626767a7
Change-Id: I12ea13555f4dac157571fdf2f5ed49a99ee3a168