[roll] Roll fuchsia [rfc] Package Sets
This RFC provides an official definition for the term "Package Set" and proposes
an updated framework for package sets. This framework expresses packages sets
in terms of the availability and version control they provide.

This was previously in work as fxrev.dev/761803 but as an external contributor
I'm not able to add new patchsets to that CL.

Original-Bug: 110948
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/794079
Original-Revision: 185a481b9364cd60f1e2feb68ab5e515bf85b2d1
GitOrigin-RevId: 86f49f569ebbe614aa8469bee4429ce17401029a
Change-Id: I919dafbb0073dd6b89734ee37d6bc31b24d8fa3f