[roll] Roll fuchsia [fxtest] Use incremental publishing by default
Integrating the incremental publisher into `fx test` has been
challenging, since it runs in parallel with the rest of the build. That
led to races between the two, and caused issues where some packages were
not published.

To fix the race, we changed `fx test` to also trigger a publish after we
did a build, to make sure all the packages we built actually end up in
the repository. This ends up removing the need for having a file-watcher
based publisher.

Because of this, this patch removes the need to have the background
incremental publisher running, by adding a feature flag specifically for
`fx test` to opt into this new workflow.

Original-Bug: 122023
Original-Fixed: 123180
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/815570
Original-Revision: 11deef190dba9cc3a449ec6d790ec5cfbe4f909f
GitOrigin-RevId: b6a3be166e51024dc529ce0b42c47a4c144f9989
Change-Id: I218ebba228c2833b241d22c7e8556d2f3567f65a