[roll] Roll fuchsia [kernel][vm] Split interval and add page
The new LookupOrAllocateCheckForInterval function can add pages in the
middle of an interval by splitting the interval around the offset that
is being looked up. If the offset does not fall in an interval, calling
this function is similar to calling LookupOrAllocate.

Not used outside of unit tests.

Test: vmpl_unittest
Original-Bug: 122842
Run-All-Tests: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/812808
Original-Revision: 073c5ac1cbf8622292d3bfb79abe83ca13aad642
GitOrigin-RevId: 953747f6f59ee6d5adfb0c1dfaf0eff65492d07c
Change-Id: I7cfccdd66b8b65daf601071ab5872adc7ab17059