[roll] Roll fuchsia [media/ffmpeg] restore required 'decoders'
Little-endian PCM decoders were dropped from the max and default
ffmpeg profiles, because we don't need those decoders. Those formats
can be consumed directly by our audio renderer. As it turns out, the
demux will not identify LE PCM streams correctly if these decoders
are not enabled. This breaks a test, so those decoders are restored
in this CL.

When we create product-specific profiles for smart displays, we won't
need to demux these formats, so we'll be able to drop those formats,
as well as most others.

Currently, the ffmpeg roller is broken, because we are using a config
that lacks these decoders. Once this CL lands, another roll will be
done, and the ffmpeg roller should succeed.

Test: build issue, manually tested
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/816074
Original-Revision: 28b42025730223540ab5c3f213f78c2f7c233715
GitOrigin-RevId: 52d6561e73df46d97baade127d4eb01a4e710968
Change-Id: I4c8736be7bacde654a6113d2c70de674915b83e7