[roll] Roll fuchsia [drivers] Migrate last users off of device_get_profile
These should likely be using deadline profiles rather than high
priority, but in an effort to remove the API it is being migrated as is.

Also, because profiles are best effort, errors in setting the profiles
were modified to be warnings.

Original-Bug: 40858
Original-Fixed: 86042
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/815687
Original-Revision: 1961c0c560bfc692f0780dd54d4d14091f4226bc
GitOrigin-RevId: 85b4c6b2f43c7e8f32068290025bb6109575dde4
Change-Id: Ibe5c1a2ecb02d6bbb9e7c41cdfed07c8b297c02a