[roll] Roll fuchsia [ffx][emu] Reland Removing manual target manipulation
This is a reland of fxr/810447. It was rolled back because
of failures in the Flutter Engine and Chromium projects.

The root cause was the waiting for the instance to start fully was relying on the old logic handling ephemeral targets had a 1 second delay built in. Removing this made the loop iterate for N times where N is the number of seconds in the timeout.

This change adds using a timer to keep track of the elapsed time explicitly.

After this change, the chromium presubmit test passed successfully.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/814450
Original-Revision: bbcd83619f49542e39e99169e8b8286a2e701733
GitOrigin-RevId: efecbe3b1cfc168335305ac6c26a00b77c1c9115
Change-Id: I27508e74d450920e2699153f62d228c7ddd54357