[roll] Roll fuchsia [starnix] Disable flaky file locking tests and add reduced testcase
The fcntl file locking tests are flaky in Starnix right now because we
do not release file locks on process termination consistently. This adds
a reduced testcase for the specific issue and marks the tests "skip" in
the expectations file to turn the bots green while we investigate a
solution.

Original-Bug: 122600
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/816568
Original-Revision: a7bfedd0a211d9978489613af6dbe5415c6ccfc3
GitOrigin-RevId: e3c18f7b7321b8f5478dc10aaf1db2439ec78911
Change-Id: I7bb49d725950b3eb9ac39f3ed0b95a0e46c9c645