[roll] Roll fuchsia [kernel][vm] Avoid gap function in LookupPagesLocked
LookupPagesLocked tries to accumulate further pages (upto 16 pages) once
it has successfully found the page being requested. It uses the
VmPagelist ForEveryPageAndGap iterator for this where the gap function
does nothing except return with ZX_ERR_STOP, stopping further traversal.
This can instead be written as a ForEveryPage iterator which avoids the
gap function altogether.

The reason for this is that when operating purely on a gap, we need to
do an IsOffsetInInterval check to make sure we're dealing with a gap and
not a page interval, which is an O(log N) walk in the page list tree. So
it is more efficient to avoid using the gap function altogether where
the gap function would have done nothing anyway, specially on a hot path
like LookupPagesLocked.

Original-Bug: 122842
Run-All-Tests: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/816490
Original-Revision: 7907babf7eaefe68456d20bccecb6844a07bdc8d
GitOrigin-RevId: 083497b983b324702ae33fb05666d01e6992a74c
Change-Id: I81045926b80489f29b93fa91220b570caba1cc07