[roll] Roll fuchsia [drivers] Add an allowlist on platform device banjo
Add an allowlist to the banjo protocol. This makes it more obvious which
devices still need to be migrated.

Remove the vast majority of direct dependencies on this banjo
protocol, because these drivers depend on device-protocol-pdev.
Some of these drivers still need to be migrated to FIDL as well.

Original-Bug: 110796
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/816887
Original-Revision: 0d84c74af86540328092082ea1eb712ec1a481b4
GitOrigin-RevId: 6464b32d28ef63be2bb9d860a0d4a9174d8b54c3
Change-Id: Ia2f25f11465b094f13b330a46083dc888573a77c