[roll] Roll fuchsia [assembly][recovery] Switch recovery images to utility feature-set-level
This should be a no-op, this switches the recovery-like images over
to using the 'utility' feature_set_level (which is currently the
same as bootstrap).  After this, and the equivalent change for OOT
images, the packages can be moved from the transitional labels
over to Assembly.

Original-Bug: 122861
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/816588
Original-Revision: 55daeccbb9112080a6a2aae42ae90a877fea230b
GitOrigin-RevId: 286a7686c9154ffe11d1c341844e86316f827d87
Change-Id: I0a7a59c2794d81386e438a1c9531ea4003131da5