[roll] Roll third_party/ffmpeg [ffmpeg] re-roll ffmpeg to 346d30b3f6631f2d47fa310f7dfeb78215706fc3
This is an update for a previous roll (fxr/807596) to the same
Chromium ffmpeg commit. Due to issues described and addressed in
fxr/816074, that roll caused some tests to fail, so the prebuilts
could not be rolled into fuchsia main.

This CL contains the results of an updated roll with fxr/816074 in
place. A manual test verifies that the problematic tests are passing
with this update. Chromium's ffmpeg repo is still on the same commit,
so there's no need to update that.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/third_party/ffmpeg/+/816031
Original-Revision: c6974898e348041595859cb30d3689b537bb61c1
GitOrigin-RevId: 81d40ee99b1844c7cb4a59ee0758098eecf70e7b
Change-Id: I6bfcf57616b2dca4df4c0147e3cffe37d5953935