[roll] Roll fuchsia [starnix] Work on ARM64 compilation.
Remove definition of MINSIGSTKSZ from signals/syscalls.rs. This is
defined in the generated uapi files so we can avoid maintaining our own
definition. The disadvantage is that the type is u32 so we now need to
cast to usize when we use it.

Correct the definition of nlink_t and use it (previously the x64 didn't
use this type because it was wrong).

Use blksize_t definition where appropriate.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/817027
Original-Revision: e19fa1eea4a9990137d07b0a88051197853a7877
GitOrigin-RevId: b0088e792060e20bdd8c88cd4c9d806c9355c8d2
Change-Id: I906896a760da1375ac0572d798226884821f77e4