[roll] Roll fuchsia [a11y] Adds Gesture Handler v2
This change is part of the migration of a11y services to the new gesture disambiguation protocols.

This change is just a copy of the old Gesturehandler, changing a few data types to match the new expected format (mostly GestureContext).

Once the migration is done, this will be renamed to be back the Gesturehandler, and exists as it is for now so that we can support both gd protocols for a while.

Original-Bug: 100030
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/801596
Original-Revision: 1680d092d16c7a1bcd6e500967ba57ee698fc738
GitOrigin-RevId: e34ea7efab67441c4fdafcedb113befb36e5fc7e
Change-Id: I63f9856f65d6b781e65bc3e0883bc8e5010b1621