[roll] Roll fuchsia [fxtest] Build `updates` if package not in publish list
The incremental publisher uses the `all_package_manifests.list` to know
which packages to incrementally publish. That file is created by `fx
build updates`, so it can get out of date if the user has changed the
build topology.

This changes `fx test` to check if the test we want to run is in the
`all_package_manifests.list`. If not, that file is probably out of date.
When that happens, run `fx build updates` so we know what to publish.

Original-Fixed: 123398
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/817251
Original-Revision: 75e319ef160c8b1125d38039cec2284444f4bb1a
GitOrigin-RevId: 748c6ec9a8f4fac31fda9c104c6885067de0dbd0
Change-Id: I079b3569b8bada8b84cd4ac5fd3dbf7773cde21b