[roll] Roll fuchsia [netstack] Add config-data for GOMAXPROCS
Add a config-data directory for netstack and use it to configure
GOMAXPROCS, which is the number of operating system threads that can be
used to run Go worker threads simultaneously. Expose the field in
inspect, as well as the number reported by zx_system_get_num_cpus.

For now, providing config-data to the netstack is optional. This is
because there exist many usages of a hermetic netstack, such as in
out-of-tree tests, that do not currently provide config-data and which
would need to provide test-specific data if it was required. Instead of
making it required, default to not modifying GOMAXPROCS. Also, make the
GOMAXPROCS field of the config-data itself optional, such that in the
default scenario, an empty configuration is provided and the netstack
uses the Go runtime's default value.

Original-Bug: 123194
Test: netstack-inspect-integration-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/815847
Size-Review: David Iserovich <diserovich@google.com>
Size-Review: Aaron Wood <aaronwood@google.com>
Original-Revision: 583a582337be2515ca39369a1a5d47f434378f1e
GitOrigin-RevId: 279ecd8e74619a9e743cb82c600e650514028480
Change-Id: Ice6c23f52d81ee8d0ccc6ed721509f37850023b6