[roll] Roll fuchsia [fidlc] Implement string_array type behind a flag
This is an implementation of the string_array type that is needed to
express kernel interfaces. It stores UTF-8 strings in-line and
zero-padded. This is gated behind the zx_c_types experimental flag,
though maybe it deserves its own flag.

It supports the syntax: string_array<N> to declare a string array type
and is represented in the JSON IR as a having kind==string_array and an
element_count reflecting the number of bytes.

Internal to fidlc it's represented as a constraint on array so that all
of the array layout logic can just flow naturally.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/816062
Original-Revision: 8557c9f0f426486a7887119cf18692513d9ded86
GitOrigin-RevId: e1afc1fa1ada42e8e0891f2d79ece56420b74cec
Change-Id: I3be9561f5559b82ccb2250d1eca11662bb6bcca0