[roll] Roll fuchsia [netstack3] Get default BufferSizes from bindings
When a new TCP socket is created, derive the default send and receive
buffers from the NonSyncCtx, which is implemented in the bindings. This
fixes a bug introduced in I4088eab8067d8843a665fe7a2b29df6ce65ab716.

Before that change, the core provided no guidance on how to size the
receive buffer, so the bindings was using the zircon socket buffer size.
When that change introduced SO_RCVBUF support, it also added a default
receive buffer size in core. The new default size was smaller than the
value previously used by core, which caused a performance regression.

This CL moves the choice of default buffer sizes (both send and receive)
into the bindings, which allows the bindings to provide the same
original values (before the regression) to core when a new socket is
created.

Original-Fixed: 123076
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/816268
Original-Revision: 92c277fd090d02e5b8e506a2713ec04865a903f0
GitOrigin-RevId: 07617d401a1a2fe6b0913c891868953e946ffe62
Change-Id: I14d290c875f440b4371fa7a8874651f559c26141