[roll] Roll fuchsia [fdf][registers] Migrate registers protocol phase1
This is phase 1 of the banjo to FIDL migration for the
fuchsia.hardware.registers protocol. This is a trampoline protocol and
there is an existing FIDL protocol by the same name. Since there are a
decent number of clients (some out of the main repo) this will be a soft
migration. That means we first have to support both.

This phase will:
 - Create a service under the FIDL for this protocol
 - Migrate the bind library's name to match the FIDL library
 - Support serving the FIDL service in the driver alongisde banjo

Tests: Client behavior is currently unchanged
Original-Bug: 123422
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/817329
Original-Revision: 3c85b5a27fa1dcc401b5dd90c49d0df6408abe5f
GitOrigin-RevId: dbe6722aebfdc50e13b2bcf715707001c66cf7ff
Change-Id: I2abcfb57fbb112728862150866c546140c66fbd4