[roll] Roll fuchsia [a11y] Instantiates Gesturehandler inside GestureManagerV2.
This change is part of the migration of a11y services to the new gesture disambiguation protocols.

The code here is just a copy of the original Gesturemanager, instantiating a different data structure (GesturehandlerV2).

This is necessary in order to support both gd protocols for a while, as we do the migration.

Original-Bug: [a11y] Adds Gesture Handler v2

This change is part of the migration of a11y services to the new gesture disambiguation protocols.

This change is just a copy of the old Gesturehandler, changing a few data types to match the new expected format (mostly GestureContext).

Once the migration is done, this will be renamed to be back the Gesturehandler, and exists as it is for now so that we can support both gd protocols for a while.

Original-Bug: 100030
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/801597
Original-Revision: 080ca4c70b86bd65704c54da9fc07395f191292d
GitOrigin-RevId: 721fb2a8fef28338bb1c6b877d15e4af822d2320
Change-Id: I936bec3c9977882ab9b23c2812da4669bd7d3552