[roll] Roll fuchsia [fdf][ddk] Remove load_firmware_async
This API was added to solve a deadlock issue in a usb driver, but was
seemingly not used to solve that problem. It's seen no adoption outside
of tests since then. There seems to be no reason to keep it around.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/816727
Original-Revision: 9992ac5b117c9ca3490d30d1460d19699ee4d273
GitOrigin-RevId: ca5bf2e1670f601ea63e35b3120ca94ac0de1059
Change-Id: Icee92572500d176de6c5c852033d4dee4b3d1531