[roll] Roll fuchsia [a11y] Accessibility view instantiates a touch source.
This change is part of the a11y migration to use the new gesture
disambiguation protocols.

This change instantiates, as part of the GFX | Flatland Accessibility
View creation a touch source, which is immediately upgraded to an
augmented touch source.

In order to support both the new API and the old one for a11y pointer
events, this changes introduces the following notion: - a11y view
continues to be created only once; - A touch source is created together
with the a11y view; - The touch source can be taken from the a11y view,
and can be returned to it when it is no longer being used; 1. This is
necessary, because the old API creates / destroys gesture arenas as
needed; 2. Once we get rid off the old API, the new gesture arena will
only take ownership of the touch source and keep it.

Original-Bug: 100030
Test: fx test a11y-lib-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/801598
Original-Revision: a7df8882d907e9828c8686831458b4f90463fde7
GitOrigin-RevId: f17195de2ce5b819e3b7b718544fb27b977a5378
Change-Id: I7061eb957e3c31f7d4ec8a906fb7b6fc68a826f3