[roll] Roll fuchsia [fvdl] Removing dependency on package_archive
The use of //build/gn:package_archive as an input to the test
is causing the creation of that archive on effectively all
incremental builds. This is a 40 second operation with a repository
of this size (this is an operation that doesn't scale with the
size of the incremental build, it's a constant-sized operation based
on the size of the overall TUF repo (base + cache + universe)

The package repository was not actually used in the test, so it is just
removed, and an empty directory is created instead.

The tool `fvdl` is being actively deprecated in favor of `ffx emu`
so the test is not actually being modified to improve coverage of
successfully starting package serving. This will be covered in
`ffx emu` e2e tests which should be written in the near future.

Original-Bug: 123378
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/817851
Original-Revision: cfb40eb8c1bea56d48a126c6a35549f999fdc3c0
GitOrigin-RevId: b73259587eda7d0760f7780255e1e19f7fc49a80
Change-Id: I33aaad44e275a56d7ce6c61e7d536d305855f766