[roll] Roll fuchsia [tracing] Add producer targeting to perfetto-bridge
This change adds the ability to target an enabled category at
a specific perfetto producer to perfetto-bridge. It's modeled after
the ffx trace feature so category names take the form of
<producer_name>/<category>. This means that categories without a
producer prefix are applied to all producers, and categories with
a producer prefix are only applied to the producer matching that
prefix.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/793129
Original-Revision: c680c4d71049089f9bc992f14b3547538ad86ac3
GitOrigin-RevId: 7aad66468b7b877f3220762fa1974890fb2f23c9
Change-Id: I24dc6b4ea125f0537694ba9c32c12e7cb2394d73