[roll] Roll fuchsia [network] Fix GN arg to enable NS2 tracing
Now that the selection of which netstack to use is a part of product
assembly, the GN arg to enable tracing which switches the netstack
package in `//src/connectivity/network:netstack` no longer works for
enabling tracing in NS2.

Fix this by having the GN arg switch out the component in the
`netstack2` package (which is selected by product assembly) for the
component with tracing enabled when `enable_netstack2_tracing` is true.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/818110
Original-Revision: d33351d16b6511a0e9ea44baeef1a19db205b807
GitOrigin-RevId: 02f265edfea5d8ab20935a2d621f94ca8252b289
Change-Id: I2217af0b958d3fd13a4a46b7a7d831971be3892b