[roll] Roll fuchsia [media][codec] Add support for LC3 codec.
Most of the parameters required for LC3 encoder is in PcmFormat struct.
For the parameters that can't be represented by DomainFormat, we add
them to Lc3EncoderSettings. For decoder, the compressed audio table is
empty since most of the information can be embedded as part of
`oob_bytes`, similar to other codecs.

Original-Bug: 122847
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/812570
Original-Revision: eb1cb7250cbe4ffda29c86a125ef5f84912365a6
GitOrigin-RevId: d1078ef9c113b1ad00fb196aa48af32f55b214a3
Change-Id: I6324692f847b3e33f9e9bf8580b373087c9e949b