[roll] Roll fuchsia [assembly][recovery-like] Transition core_realm and core_realm_networking packages to assembly
Split the 'core_realm' AIB into two pieces, with the
AIB containing all '/core' children that are expected to be present when
the '/core' realm is present, and networking is present.

Change all packages in those two AIBs to be provided by Assembly, and
not the transitional labels that are being used to add them to the
recovery-like images.

Original-Bug: 123397
Original-Bug: 122862
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/817230
Original-Revision: dc4d113dc3adab4c1c34ab2e490c5bef8c6a58bc
GitOrigin-RevId: d7bc2b4fef9cb73737828c8df612c9d4cbeb4d15
Change-Id: Ied45048d70d824299eba4599c41c1de5dbd7513d