[roll] Roll fuchsia [display][core] Use a custom trait for IdMappables.
IdMappable is a trait publicly inheriting SinglyLinkedListable,
which implements methods that required for SinglyLinkedLists like
InContainer(). This makes it possible for the IdMappable child
classes to be included in a IdMappable::Map, but this way we cannot
use the IdMappable class in any other intrusive containers.

In display core we work around this by defining a "linked_list_node"
within the type and let it refer to itself and get linked to a
linked list. Today, Image and Layer use such self-referencing nodes.
This is risky and if used improperly could cause memory leak
(fxbug.dev/123301).

This change replaces the existing IdMappable class which inherits
SinglyLinkedListable with a NodeState class composed within the
class. The linked list is created from SinglyLinkedListCustomTraits
instead. This way we allow IdMappable nodes to be able to exist
in other intrusive containers, while not introducing any extra
allocation overhead.

Performance-wise this should be totally equivalent to the previous
IdMappable class, without any extra time / memory usage.

One drawback is that IdMappable won't support SinglyLinkedListable
methods like "InContainer()". Since these methods are not used
in today's IdMaps, and we don't move Id-mapped resources once it's
created, being unable to use such methods shouldn't be a problem
for display core.

Test: display-core-unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/816229
Original-Revision: 32d1e88a32cf1db911fb4cf2b1701f6f0b312eb5
GitOrigin-RevId: b3c9b0224d7b2bd16345fa72c2670b0e478ddf7b
Change-Id: I652868722f7c96308e1a667ef29c987842dcc851