[roll] Roll fuchsia [radar-proxy][test] Make threading consistent with real implementation
Make the test only access the radar proxy object and its dependencies on
the dispatcher thread, like the real implementation does. This is an
attempt to fix a rare flake that occurs in CI once a month or so.

Multiply: radar-proxy-test: 1000
Original-Bug: 123502
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/813642
Original-Revision: 76fc5f25d5e50a42a145fcb5e8547bb16f002771
GitOrigin-RevId: f87ebd5fc5e39e525e9f9d9adecafa2687bb4e55
Change-Id: I20460e1bb3f6ec9d933e8f8ae27f2a028bdf3a41