[roll] Roll fuchsia [kernel] Distinguish __executable_start and __code_start
Some existing uses of __code_start have the purpose of finding
the beginning of the kernel load image overall.  Other existing
uses have the purpose of finding the start of the executable
segment.  In the future ELF layout these should not be the same
thing.  So this changes the existing kernel linker script to
define __executable_start in addition to __code_start, and
changes references that are for the "load image" purpose to use
it instead.  The names are unfortunately a bit confusing--the
"executable" in __executable_start sounds like it might mean "the
executable segment", which is the very same thing as "the code
segment".  However, it actually refers to "the executable" as in
"the executable file", or "the binary".  But __executable_start
is the symbol name that the linkers actually provide by default
for the "load image" meaning, so we stick with that name.

Original-Bug: 122872
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/816567
Original-Revision: e27012adf4b85b8a8e2deb8bb1de884d6e019af4
GitOrigin-RevId: 13fad91e46913deaea9c688c82f2298926d1a646
Change-Id: I2a31cee3586309ac9d79b1c8c1262ea5b2b719d9