[roll] Roll fuchsia [ffx][repo] Return target communication failure if target offline
If ffx loses connection to a target right when we're trying to register
a repository, we might get a `None` for the target nodename or ssh
address if ffx has only partially torn down the connection to the device.
If we see this happen, return an `TargetCommunicationFailures` instead
of `InternalError`, since this doesn't appear to be an internal bug.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/819187
Original-Revision: f5c8b185d11102f879b965ad8cf37f411677cb67
GitOrigin-RevId: 4f32c6b526accf343436789aa40b0547ab1a5305
Change-Id: Ia62818b0930bae38ee95d03b01f2eb047d3215a0