[roll] Roll fuchsia [netstack3] Plumb route metrics through ns3 core
This CL introduces the concept of a route metric to NS3 core. The metric
is set when creating a route, and can be later read back when listing
routes, or observing routing events.

Note that this CL does not introduce any changes to how routing
decisions are made. Honoring the route metric when sorting the routing
table will be introduced in a future change.

Original-Bug: 122489
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/810928
Original-Revision: d2345de24ee0d4dbd57de55944fa5e5987d6568b
GitOrigin-RevId: 77d5bc4911b50afdbbbeb277da37903c5ea1ca67
Change-Id: I6505ee16d7919338fd7fedb503058b9c8bbbb3bd