[roll] Roll fuchsia [pkgdir] Refactor tests to use a test realm proxy
- Replace test.fidl.pkg.Harness with fuchsia.testing.harness.RealmProxy
  which will eventually be partner-facing in the SDK.
- Reimplement the pkg-harness component as a test realm proxy.

The new test realm proxy doesn't support switching between the
"backing" of the package directory; it looks like this feature is
not required by the tests anymore anyway. If it is needed in the
future it can be accomplished by subpacking the test with a
different proxy implementation that uses the new backing, or a
test harness that spawns RealmProxy instances given some object
specifying which package_directory backing to use.

Original-Bug: 121395
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/794462
Original-Revision: 9c0d3e9cfde58fe824942997009493c6b3dbebba
GitOrigin-RevId: 7d0e66cc43f5bd2494c2321ff3007be1d3da700c
Change-Id: I7b11d37c7c81cea98478f518b0e1f405553ad206