[roll] Roll third_party/pigweed pw_stream: SocketServer: allow reusing the port
Previously if the server started listening on port 0 it would not use
SO_REUSEADDR, so if later on the server was restarted on the same
(resolved) port number it would fail to bind. In order to fix we are
moving setsocktopt() call so that it happens for port 0 as well.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/133991
Original-Revision: c2df1e4942303b0e91f753f2e68b944b96fa9a93
GitOrigin-RevId: 22b9528bd2c5b37fab6cfcb4ba01ace3fcda9cd9
Change-Id: I50b50893db157879c388a23401476a9a990be90b