[roll] Roll fuchsia [superproject] Roll third_party/pigweed pw_stream: SocketServer: allow reusing the port
Previously if the server started listening on port 0 it would not use
SO_REUSEADDR, so if later on the server was restarted on the same
(resolved) port number it would fail to bind. In order to fix we are
moving setsocktopt() call so that it happens for port 0 as well.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/133991
Original-Revision: c2df1e4942303b0e91f753f2e68b944b96fa9a93
GitOrigin-RevId: d11f7b37fad4f25f02bb129e6988307ad7aad4bf
Roller-URL: https://ci.chromium.org/b/8786540421932659649
CQ-Do-Not-Cancel-Tryjobs: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/819830
Original-Revision: daf81efc95b9625f91b34959302ca6c427d1db36
Change-Id: I359911ffa07dfeccf2b4a9c164d9acffb724595f