[roll] Roll fuchsia [starnix] Release the file descriptor of a Task at the time it dies
Currently, the fds are released at the time the Task id dropped, but
there is some case where the Task survives (because it is in a Arc)
while it is already dead. This ensures the fds can be released at the
right time.

This should be unnecessary when http://fxb/122600 is resolved.

Original-Bug: 122600
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/819807
Original-Revision: d1c370692a7e5d58b5cf9d3edc9aa3653feca434
GitOrigin-RevId: 1a4f32ecf4ac69bbe80cb84dbcc273b1a8fa635d
Change-Id: I90eada704256628e5dec47960ff2ebff3052c67a