[roll] Roll fuchsia [ffx] Generate test data for ffx and subtools
This makes it so that when we build a subtool, we add a target that
generates 'test data' for it and puts it somewhere we can get at it.
This allows tests to not reproduce this effort every time they want to
call ffx tools.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/812172
Original-Revision: e630de7062135baaac72228d94ac8a000fd043da
GitOrigin-RevId: b54f5df7f95220d8a1ac99f4e9bd6b677ccb1b2a
Change-Id: Id6a7dbdfd9b3b2ffc818cf116d79676feb0ba71f