[roll] Roll fuchsia [kernel][phys] Plumb arch-specific data from physload
This moves some state from per-arch global variables into a
per-arch `struct ArchPhys`. Each ArchSetUp() implementation is
responsible for setting `gArchPhys` to point at an initialized
instance of this (if it's used at all, not so far on x86).  This
is plumbed through physload handoff to set `gArchPhys` in the
physload module to point to the one set up earlier.

On arm64, this includes the PSCI information which heretofore was
being lost on physload handoff so that physboot's uses of PSCI
didn't have the state set up properly. (This only affected the
panic path. ArchPhysHandoff doesn't use these globals.)

On riscv64, this provides a way to plumb the boot hart ID,
which is now included in ArchPhysHandoff.

Original-Bug: 122871
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/819530
Original-Revision: 662eb9247c8e43070f1d3443c79ed30cada0827b
GitOrigin-RevId: e79b9d10480bf74a98167dbd911ef85deacc5a71
Change-Id: Ibd667cc5a62f2e71d27899401d63c1e12fcb9c15