[roll] Roll fuchsia [ffx] Clean up host ids in RCS
When a new connection comes up, the remote-control-runner calls
AddHost (legacy) or AddOvernetLink (CSO), adding a new u64 id which
gets appended to a list. After 8K of these or so, we can no
longer respond to IdentifyHost calls because the DISLmessage is
too big.

The fix is to have each client loop maintain an Rc to each of its
ids, and the handler only send back the ids that have not been dropped.
So when the RCR exits, the loop will exit and we will no longer report
its ids.

Original-Fixed: fxb/123624
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/819297
Original-Revision: 6d806643034983b37568868d7bd214abc78b8895
GitOrigin-RevId: a21ed1c63a68c635bc79a5c75f1b9c65bce8a1ea
Change-Id: I8181298b97abcb33ce2edcd36a48b0b73acfac81