[roll] Roll fuchsia [fidl][cpp] Client side sticky errors
This is the client side counterpart to
I35eed7f60f7f82ed060d3af487f21f7f5d22394b.

The client goes through two indirections to obtain the transport.
Firstly the ClientBase ref-counts an AsyncClientBinding. The
AsyncClientBinding ref-counts the transport. This is because there might
be other sync calls that will want to also hold on to the transport.

Thus the logic on the client side is a little bit more involved but
we're just writing variant manipulation code to simulate the kind of
monadic Result operations found in Rust.

Added unit tests since this is unlikely to be general enough to warrant
a dynsuite test.

Original-Fixed: 87788
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/819300
Original-Revision: 0827532d56522af2877a635db2f6a02dbf20ec6b
GitOrigin-RevId: 865cf252abfcb4d2d202e30989a01822aeb7cbff
Change-Id: I52c9b2215b9fccd6cc0efb3c126ac6acfade7b7f