[roll] Roll fuchsia [magma][msd-vsi-vip] Fix WriteUnorderedEventIds
Reading and clearing the interrupt failed if the NNA was suspended.

The test 'WriteUnorderedEventIds' bypasses SubmitCommandBuffer. This can
result in the device suspending in the middle of the test.

Added an option to DeviceThread to disable suspend.
This option is for use in testing.

Change logic to for hangcheck timeout to ensure suspend timeout cannot
trigger a hang.

Change debug messages in interrupt routine.

tested: msd-vsi-vip-integration-test
bug: fxbug.dev/123433
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/819189
Original-Revision: fb8bdd1ba01d805c9a4da7d4ba0fef40db6c36be
GitOrigin-RevId: f986a9cd20eb53361622b6ea5a8875fcc5295f3e
Change-Id: I2b1ec878c8d144dd62ee85050e2d7b22647b2807