[roll] Roll fuchsia [netstack/bench] Report lost packet %
Include the percentage of packets that were lost (as opposed to just the
total count) in the results of the iPerf benchmarks. Also, mark the
"lost_packets" metrics as "smaller is better" so that decreases are
correctly detected as improvements and vice versa.

Original-Fixed: 123437
Test: catapult_converter_unittest
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/819788
Original-Revision: 88b114dfa6f6bea78f6479107ad3183dd104bbc8
GitOrigin-RevId: fb26d5326af54774909d8cd627d2366a77d8f821
Change-Id: I8178026b0a5f245f7819b1946315099b211d722a