[roll] Roll fuchsia [fake-display] Make fake-display not a system driver anymore.
fake-display should be only allowed to use within a fake
device tree provided by fake-hardware-display-controller-provider
in a test realm (as specified in Fuchsia RFC-0180: Test UI Stack).
It shouldn't be used as a system driver in a non-hermetic environment.

This removes the driver binding logic, the driver component and the
fake display "device" specification so that fake-display will not be
able to bind to any device.

Test: core.qemu-arm64 builds
Original-Fixed: 123816
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/820834
Original-Revision: 5e36fd907013d373699db40e11a24cbf2d43a990
GitOrigin-RevId: add44be6c2ebb856f2a63bcca007c6f119850b96
Change-Id: Ie383c735f2d5e4617e15bac43a9d98922c54f237