[roll] Roll fuchsia [benchmarks] Split tracing benchmarks out from microbenchmarks_test
One reason for doing this is to allow flakes in these sets of tests to
be tracked separately (by flake-fetcher and other tools), because we
have found a couple of flaky failure modes that affect the tracing
tests recently.

This also has the benefit of making these tests a little more
convenient to run locally:

 * tracing_microbenchmarks_test is much faster to run on its own.

 * microbenchmarks_test doesn't depend on the SL4F server now
   (following the recent change to runTestComponent()).

Note that we just moved microbenchmarks_test.dart to
tracing_microbenchmarks_test.dart and removed a small part, but
Git/Gerrit do not recognise that move.

Original-Bug: 123479
Test: fx set terminal.x64 --release
  + fx test --e2e host_x64/tracing_microbenchmarks_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/819210
Original-Revision: 46f7efa1c48beb86b5cb782ca18ae22c6ff67700
GitOrigin-RevId: ee38f14d66df6f22dc7ab831f858a6978d5d1b9f
Change-Id: I155c0bee1ea3f73dc2096256ebb98fa346fb642f