[roll] Roll fuchsia [netstack] Enable inline static init optimization
This was disabled in http://go.dev/cl/467015 due to bugs that cause
compilation failures. As we are not affected by these bugs, it's safe to
turn this on and avoid the 300KiB bloat that is otherwise caused by
upgrading Go from 1.20 to 1.20.2.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/821088
Original-Revision: d316b8e61489dd8053077e251d002f274985a302
GitOrigin-RevId: a54663a1ccea624ac2455355f18766ba0bebac1a
Change-Id: I905e11ffa103006807b4a65e86d1f6827a53abc8