[roll] Roll fuchsia [overnet] Fix race in ServiceMap
ServiceMap was trying to be clever and have different locks for
different data, but the data had to be kept in sync and it wasn't quite
clever enough to pull that part off. This removes a bunch of the
interior locks and replaces them with one big lock around everything.

A side effect of this is that a couple of functions that didn't have to
take that lock now do, so they become async.

Original-Fixed: 110501
Original-Fixed: 122211
Original-Fixed: 113099
Original-Fixed: 98883
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/821128
Original-Revision: d84f365acb67cbcee8b11ad17c5766cd1f3e7da6
GitOrigin-RevId: fd8eaa11128b84b3e5db1dfbda680570fe6c1432
Change-Id: I1fb65195f20ab29bac6596074b5681885796b2b6