[roll] Roll fuchsia [display][core] Replace DoublyLinkedListNode with fbl NodeState.
This replaces the previous DoublyLinkedListNode which contains
a RefPtr to the Image itself, with fbl NodeState class, which
encapsulates all the RefPtr manipulation, making it safer for
developers to use.

Note that since the node has to be mutex-guarded, we cannot
directly use DoublyLinkedListable traits, instead we have to
add a mutex-guarded DoublyLinkedListNode member to Image. The
implementation is the same, but the methods are checked by
compile-time thread-safety checker.

Test: display-core-unittests
Original-Fixed: 123301
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/818351
Original-Revision: 863fd29ffcbb1aec8d8dccf54402ee98a11b9779
GitOrigin-RevId: b187254fb4eff3ad6196370d8d6eacbea2176618
Change-Id: I99520033293c5aedebec43483dacd3eac4a4c5b8