[roll] Roll fuchsia [starnix] Do not swallow binder errors
Binder errors are currently, in the vast majority, an issue with Starnix
code. Because of this, and to make progress, they were currently
swallowed, and a success was returned to the client code.

The starnix binder code is now in a better situation than previously,
and the user mode code has been modified to crash loudly in case of
binder error.

It is now beneficial to see these crashes to be able to correlate client
code and binder code, so the error swallowing mechanism has been
removed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/821789
Original-Revision: d5d9371a91152dd5af83a0aa77369428a5faa524
GitOrigin-RevId: 1785f73058db1234e332e5d5e2dc45320f2ce1bf
Change-Id: If3ced6f617ecd2cfe96c6700a46ef6d4cb8a7eb7