[roll] Roll fuchsia [fshost] Don't assume Block/Volume multiplex
Refactor so BlockDevice retains a ControllerProxy and vends new
Block/Volume connections via ConnectToDeviceFidl.

Filesystem also requires a ControllerProxy, so BlockDevice now has a
method to reopen a ControllerProxy.  Internally this uses the path
stored in the BlockDevice which is not ideal, but we can change things
around (e.g. to have BlockDevice hold onto a DirectoryProxy containing
the service node) if we care to do so.

Original-Bug: 112484
Original-Fixed: 120984
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/818690
Original-Revision: 3856a3b8861eb29e665f3d148919f17e486de104
GitOrigin-RevId: fd332fa75406a5320ab6b32010cd374e46b872b8
Change-Id: Ib86a1733e4f7fcd139595bfdd01ac3bfe5d9b228