[roll] Roll fuchsia [ctf] Add Subpackaged gtest_runner Shard
Adds a shard to include the gtest-runner and lib_loader_runner_cache
(a dependency of all test runners) as subpackages. This shard defines an
environment where the subpackaged test runner exists. Tests using this
runner should start the test component in this environment. This will
enable gtests in the CTF by ensuring that the gtest runner is
compatible with the test.

Original-Bug: 123826

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/821231
Original-Revision: bae04c69e41e9caf89ee72bbb2089b64ebe3652e
GitOrigin-RevId: bbedee7653028a819ee14f29fffcae58dc1c296f
Change-Id: Ie81dac2a089bc1e4776d0d295feec46dea0aa006