[roll] Roll infra/recipes [checkout] Remove min_rebase_distance feature
`min_rebase_distance` was an experiment to try to improve the
incremental cache hit rate by trying to skip fast-forwarding the cached
checkout to HEAD as long as it wasn't too out-of-date, so the patchset
under test would get rebased on the cached revision rather than on the
current HEAD.

This feature was never productionized and is still contentious (due to
the increased risk of conflicting changes being able to land). So it's
unlikely to be used any time soon, and can be deleted.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/821791
Original-Revision: 819bf619c3de634dd152d2b4b1827d422dc44f16
GitOrigin-RevId: 0997cddc6222f52b4e193625c39695698caaa252
Change-Id: If13060253ab7b8c61b6edce28014b7402f301867