[roll] Roll fuchsia [starnix] Handle forking of writable pager backed VMOs
Prior to this CL, we'd always clone if the VMO was pager backed, but
that's the wrong thing to do if the VMO was writable.

Unfortunately, we can't (yet) snapshot pager-backed VMOs, so we
workaround that (for now) by copying the VMO.

This should have limited impact currently because there should be few,
if any, cases where we use pager-backed VMOs, but we will have them
soon.

Original-Bug: 123742
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/820331
Original-Revision: bbf005c336c2e0db84f07bf3d51ecb3f70952b7d
GitOrigin-RevId: 1f8fd4378be023b6ef70de4010c1ca9f9827cb4b
Change-Id: Iaeaf1b8279bae2eaeab5a21d9ca3fa5dbc1442e4