[roll] Roll fuchsia [starnix] Fix signal stack underflow
When setting up the signal stack, it was possible for the stack
computation to underflow, causing a panic in the rust code. The code now
does checked arithmetic and handles the error.

The error is handled by delivering a SIGSEGV signal to the application.
If SIGSEGV is masked, it will be unmasked; if it is ignored, the signal
disposition will be reset to default before delivering the signal. If
the application has a handler for SIGSEGV, the signal will be delivered
to the handler, and if it fails, the signal disposition will be reset to
default and the signal redelivered.

Test: fx test syscalls_test
Co-Authored-By: marinaciocea@google.com
Original-Bug: 118658
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/802168
Original-Revision: 84d845ecce84a6d6d080d9957245ff6367ac7b0c
GitOrigin-RevId: 3e747175a3ba9d2f7d0fccee6bb1e61c7c1910cb
Change-Id: Id0508e8130aeea3dfb9197e5c31de5f5ccc902aa