[roll] Roll third_party/pigweed pw_i2c: make Device objects movable
Currently the Device object is not copyable or movable. This change
adds back the default move constructor. The Device is just a wrapper
around an address and Initiator reference, so it is safe to move.

Add documentation about the concurrency requirements of Initiator and
Device. Explain that Device is intended to represent ownership, and
that's why it is not copyable.

This change is intended to make it easier to construct objects which
depend on a Device. Previously the Device had to be allocated separately
and passed by reference to the constructor. This change allows using
temporary Device objects when constructing higher level objects.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/134610
Original-Revision: 6e6232b7777cfa7f40aaccf27ed7df53da08a0bf
GitOrigin-RevId: 45a0efc65da7ed993b292609400f0a5bc8413898
Change-Id: Ie2145d4624925645a937c3d63150403ddd9f5e15