[roll] Roll fuchsia [superproject] Roll third_party/pigweed pw_i2c: make Device objects movable
Currently the Device object is not copyable or movable. This change
adds back the default move constructor. The Device is just a wrapper
around an address and Initiator reference, so it is safe to move.

Add documentation about the concurrency requirements of Initiator and
Device. Explain that Device is intended to represent ownership, and
that's why it is not copyable.

This change is intended to make it easier to construct objects which
depend on a Device. Previously the Device had to be allocated separately
and passed by reference to the constructor. This change allows using
temporary Device objects when constructing higher level objects.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/134610
Original-Revision: 6e6232b7777cfa7f40aaccf27ed7df53da08a0bf
GitOrigin-RevId: 56266cf279f16998bed6ddab4ab2b640f6988d25
Roller-URL: https://ci.chromium.org/b/8785998850451302657
CQ-Do-Not-Cancel-Tryjobs: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/822707
Original-Revision: aa0d6054614e0824c84b7dad5de4c9c5dc7fc7cc
Change-Id: I82cf34293f6c63c490ce0bae67bd3118d99b89f4