[roll] Roll fuchsia [fxfs] Unwrap keys in the background
Unwrapping keys in the background when opening a file significantly
reduces the time it takes to respond to open calls. The keys are only
necessary for encrypting and decrypting the contents of a file. The keys
will almost always be ready before the contents of the file need to be
encrypted or decrypted.

Original-Fixed: 123394
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/818589
Original-Revision: c28e6f5c8c503a8376a53ce8ae7db2cb8ad14a5a
GitOrigin-RevId: 85f94f31fac65b1d0ba5989f32c55a7c7544d715
Change-Id: Iebbef9f9fb7f3d7ae5774171e4bf58e4dac52149