[roll] Roll fuchsia [fbl] Support non-default-constructible types in fbl::static_vector.
The fbl::static_vector documentation states that it's modeled after the
proposal in [1]. The motivation section states that "The static_vector
container is useful when:" [...] "std::array is not an option, e.g., if
non-default constructible objects must be stored".

This CL tweaks pop_back() and clear(), so they compile with
non-default-constructible types. clear() is particularly impactful,
because is indirectly used in many other fbl::static_vector methods.
This fix makes fbl::static_vector usable in at least one scenario that
involves a non-default-constructible type.

This CL also adds tests for pop_back() and clear(), so the fixes stick.

[1] https://www.open-std.org/jtc1/sc22/wg21/docs/papers/2020/p0843r4.html

Test: fx test fbl-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/822231
Original-Revision: 65a5222498135faf6e93359c25671ec000298de9
GitOrigin-RevId: 5763719fcb61b667745b219d69a651f11d1688d0
Change-Id: I4e7209123dbd26913edadc82a545a48d2f668473