[roll] Roll fuchsia [trace_manager] Add a Timeout to List Categories
Unlike Start and Stop, list categories waits for a response from the
trace provider. If a trace provider does not respond, this will cause
trace_manager to wait forever for it.

This patch races each GetKnownCategories response against a timeout, and
defaults the response to and empty list if the trace provider does not
respond it time.

Original-Bug: 123170
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/821172
Original-Revision: f84f7634ff2d57672831dc9f8ff44beab67ec4c7
GitOrigin-RevId: 5310b54188636fea1c646825d4b68853ab0df96b
Change-Id: Ibf3383d2dc7b3792f67b4d4b5b5fa911a5c13f92