[roll] Roll fuchsia [starnix] Release the file descriptor of a Task at the time it dies
Currently, the fds are released at the time the Task id dropped, but
there is some case where the Task survives (because it is in a Arc)
while it is already dead. This ensures the fds can be released at the
right time.

This should be unnecessary when http://fxb/122600 is resolved.

This is a reland of https://fuchsia-review.googlesource.com/c/fuchsia/+/819807
The changes are:
- The file descriptors are releases after the task is removed from the
  ThreadGroup to prevent races due to closed files.
- Use a named variable to force the ordering of dropped.

Original-Bug: 122600
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/821888
Original-Revision: 13b59c31b474619c7797f5d55a956210ad92dac5
GitOrigin-RevId: caeb1087c6eeb493e3c659314189e27b0a40052a
Change-Id: Ifc1956f7a8343eaba56f095f23213471a1ad59fc