[roll] Roll fuchsia [netstack3] Drop or use forgotten struct's fields
`Inner` is forgotten (`core::mem::forget`) so the destructor for its
fields will not be run. This is okay for primitive types but not for
types that may have custom destructors.

Before this change, `Inner::marked_for_destruction`'s destructor was
not explicitly called since it was an `AtomicBool` which does not do
any special work when it is dropped. However, to be safe and guard
against future changes that _may_ change `marked_for_destruction` to
a type that _does_ perform work on destruction, this change explicitly
drops `marked_for_destruction` when it is no longer used.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/821876
Original-Revision: 3821ec5b2733e3eb98cafe6c84bae63abc0f27ae
GitOrigin-RevId: 09550d48e142fd7a15ea360823d9246f86c134d7
Change-Id: I5062eefff9ac9f2eeec12e66f7cc854152a8f77f