[roll] Roll fuchsia [devshell][ffx] Build ffx subtools as fx host tools
This makes it so that if you have
`ffx_build_dual_mode_plugins_as_subtools` set in your build, if fx is
included in your build so will the subtools that are supposed to be
built. This gets `fx build` building them as it does with ffx.

Original-Fixed: 123543
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/819188
Original-Revision: 399fcbb871f0c53fefec8f40195e61f364a921ea
GitOrigin-RevId: 27bc89a31a75f37c4a2ae580992af68a8c647196
Change-Id: I4a6d3ad9ed061cffc8d478576319097a2631d4e1