[roll] Roll fuchsia [kernel][x86][boot-shim] Add a IKCONFIG to the linuxboot boot-shim
Some emulators use extract-ikconfig to determine the architecture of the
boot-shim, therefore we must insert a very minimal IKCONFIG into the
boot-shim.

extract-ikconfig greps for IKCFG_ST to find the config in the kernel.

ikconfig.txt is the contents of the config, and it is gzipped to produce
ikconfig.gz. Because this file is so small we decided to check it in
instead of adding a gzip tool.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/812288
Original-Revision: 1251720fa5eecc1237b525d480ade8fa7028570c
GitOrigin-RevId: 3cde1bbe1b560b332913ad55fd8564d5b899981d
Change-Id: I868f1507f8ff255b5cafb47cc12f9548b174d697