[roll] Roll fuchsia [fdf] Use FIDL platform device for aml-nna
Make the aml-nna driver use the FIDL version of the platform device
protocol instead of the banjo version. Using PDev banjo depends on banjo
proxying which is not supported in DFv2.

Aml-nna was updated to expose the pdev banjo protocol which it proxies
to the underlying pdev FIDL protocol. This was done to support the
aml-nna's child devices which still use banjo.

Original-Bug: 123400
Original-Bug: 110796
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/821809
Original-Revision: 3db322f17c17513095e4dcb06097863009aac76a
GitOrigin-RevId: 0d41beb165f72efc4aef32a608dd4e18e0320232
Change-Id: Ib00ab1cc6003f099cd06b99908a1250277b03c88