[roll] Roll fuchsia [kernel][vm] Populate slots in interval
Add a helper to populate slots in a sparse interval over a specified
range. After the operation, the interval will be split around the range
to: an interval before the range, an interval after the range, and
individual slots for all offsets within the range. This allows more
efficiently pre-flighting a range in an interval we are going to add
pages to, instead of repeated calls to LookupOrAllocateCheckForInterval
for each offset.

Test: vmpl_unittest
Original-Bug: 122842
Run-All-Tests: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/812809
Original-Revision: 62747079e5b587ee2bd619a942f0255d87d3e400
GitOrigin-RevId: 359060534a4e0900b314cf0509fc8f3fd0728c95
Change-Id: Ic479b2b962d837ae832848b95adc8e3c71eca686