[roll] Roll fuchsia [doc-checker] Handle +log in the git branch URLs
doc-checker detects git URLs to HEAD/docs, since these should be file
paths. In this case, the URL was ..../+log/.... and the code was looking
for either /+/ or /+show/. This makes it look for anything that begins
with '+'.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/823567
Original-Revision: bc42db09331f3d52ce945ff55a965b119a1c4f42
GitOrigin-RevId: f1e6dff5aa83ea8dde100bbf29d13cc52e61c65e
Change-Id: I55b55f209256f517919d77ac2a7953c827cf1bf0