[roll] Roll fuchsia [devfs] Expose dev-builtin (/dev/{null,zero})
This is an alternative to exposing /dev/{null,zero} as protocols which
abuses the fact that protocols are routed as RW, and to using
dev-topological which grants too much access.

The dev-{null,zero} protocol capabilities will be removed in a
subsequent change after soft transition in third_party/go.

Original-Bug: 101092
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/823408
Original-Revision: 89f9dc0747300482335a0596d04feb1b0381cbb9
GitOrigin-RevId: ca608aba211f11dd7f49ea162b02531c6a8db925
Change-Id: Iabb549974e5c6ded94911602b297ad169b8228ad