[roll] Roll fuchsia [timekeeper] Makes timesource child destruction conditional
`Realm::DestroyChild` isn't quite idempotent: it produces log spam
if we attempt to destroy a child that does not exist.

The old time source code would attempt to get to a 'clean slate' by
trying to destroy the time source component that it's about to start
first. It did so unconditionally, creating log spam in the process.

The new code first checks if there is a matching child component to
destroy. This is not error-proof, since the child component could
still go away between the time we check if it is there, and the
time we destroy it. But this should either be rare, or would point
to a bug.

Original-Bug: 124043
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/822778
Original-Revision: 6f424d6f72009aebbb8fb32f9abb1f4bdfde4855
GitOrigin-RevId: 4486362a278ac68512f7517ff08dc25e4b620600
Change-Id: Ia2cb11e6941b8412520e5d410521c8a8d3f52a1d