[roll] Roll fuchsia [starnix] Don't eagerly copy the whole VMO
It was a bad assumption that most of the VMO would be mapped. It turns
out that the mappings are often small compared to the size of the VMO,
so copying the whole VMO meant a performance and memory burden.

This change now means that we only copy the parts that are mapped.

Original-Bug: 123742
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/822478
Original-Revision: 3b8425139e673343dad1e7b89eb6c60092783609
GitOrigin-RevId: 3ce23b00f5815a6671e364e26efa9e235bf6379d
Change-Id: If8d0783c9ca3a675c7ffd5e8a8ea5058e048389e