[roll] Roll llvm-project/libc [libc] Adjust NVPTX startup code
Summary:
The startup code needs to include the environment pointer so we add this
to the arguments. Also we need to ensure that the `crt1.o` file is made
with `-fgpu-rdc` set so we can actually use it without undefined
reference errors.
GitOrigin-RevId: 9353e7a71cd35c7472c2f853d6609cf9ce4dd94a
Original-Revision: be3c04aaf110179bd44eaf8455dd5e7b42289dd5
Change-Id: I1e267fb9bf703b0608c465f34de59c4869bae19a