[roll] Roll cobalt-registry [network_policy] Improve metric comment
The metric for `internet_available_gateway_lost` was initially defined
to catch a situation we didn't expect to observe. Upon looking at
reports, it's clear that we actually see this metric.

The act of crafting an ICMP Echo packet and placing it for delivery on a
network can influence the state of that network. This metric is
effectively measuring a case where we've observed a state where there
appeared to be no network connectivity, and the host transitions back to
full connectivity.

Original-Fixed: 124184
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt-registry/+/823427
Privacy-Approval: Zach Bush <zmbush@google.com>
Original-Revision: 8d266df474dbe476a14d750554a32cd45df892b6
GitOrigin-RevId: 35f991fc1e7286b3c0651f97d55f04016cf26809
Change-Id: I848eefa7b37c4c6e8bbe300642642fde355b63cf