[roll] Roll fuchsia [fdf][compat] Add support for calling suspend during prepare_stop hook
The protocol driver_manager uses to determine the reason it's stop hook
was invoked is routed to all drivers. This mechanism necessarily must be
out of band of the component stop hook as part of the component
lifecycle.

Additionally, the Stop hook was modified to allow you to override
behavior for deleting the node. This allows us to avoid calling release
in the suspend path, which is what DFv1 does.

Original-Bug: 97457
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/744579
Original-Revision: c4211d025d032f056396dbb45dcfb349a2437d6f
GitOrigin-RevId: aa02e29045702f68cf2b68d7e10ba66994d6f9db
Change-Id: Ia02f0d7056c1a4098f848ae325c71983826a2ab5