[roll] Roll fuchsia [devfs] Add devfs-with-pkg component
There are two stages for shutting down drivers, one for boot-drivers,
and another for pkg-drivers. `/bootstrap/devfs` provides the signal for
`/bootstrap/boot-drivers` to be shutdown and the newly added
`/bootstrap/devfs-with-pkg` provides the signal for
`/boostrap/pkg-drivers` to be shutdown. This is necessary because we
cannot add a fake dependency from pkg-drivers to devfs as doing so
would cause a dependency loop when things like fshost try to rely on
devfs. This is why `fshost` still uses `/dev` from devfs after this
change but almost everyone else uses `/dev` from `devfs-with-pkg`.

Original-Bug: 108571
Original-Bug: 108572
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/818355
Original-Revision: 039a8efeba9cad20b08457fe8a88c0e04a3b62db
GitOrigin-RevId: 496b90228cbb3ae5f165c0d20bc1104744432c0e
Change-Id: I2299d96f7ce052cd5bfce09d68157a05292a6ce6