[roll] Roll fuchsia [f2fs][virt] Add explicit bash dependency.
We may offer a different shell that doesn't support "set +H" in the
VM based tests. To be compatible with that change we explicitly use bash
for the operations that depend on it.

Once we migrate to another shell (ex: dash), I believe we can drop both
bash and set +H here.

Original-Bug: 123888
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/824479
Original-Revision: 5274fa77f15e22308c19f8cf6bcb36660dfdeb24
GitOrigin-RevId: 5c215ca54dc5ff59afb893c04b670f187da5ab9a
Change-Id: Ifdad83c1006bd73189c1cfc27ca3a7ebb4433f64