[roll] Roll fuchsia [rust-3p] Update polling to 2.4.0
This updates pollng to 2.4.0, the latest version that doesn't depend
on a more recent version of windows-sys. As part of this, polling switched
to a different crate for wepoll, so this switches our empty version of
wepoll-sys to wepoll-ffi.

Original-Bug: 123569
Original-Bug: 120661
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/822968
Original-Revision: c4ba21c7bea497af1a988d4f83988c825cd6b6d1
GitOrigin-RevId: f88bd160872a567666988ff0307f00d01cb35291
Change-Id: I28fa8e47e463a28bd6117328889bc15d4f361f9f