[roll] Roll fuchsia [fdf] Switch SysSuspender to zircon transport and dedicated loop
In DFv2 the default dispatcher is shutdown after suspend is called. The
x86 board driver is suspended ahead of the platform bus driver. Thus, it
cannot use an fdf dispatcher and must use a dedicated async loop of it's
own control to listen for the callback. This also means it must no
longer use the driver transport for the SysSuspend protocol.

Original-Bug: 108572
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/824485
Original-Revision: 141f72e4138ce3b066635583f7e8768a08243f3d
GitOrigin-RevId: f69d2662bf289c5c0e4a14ed14e3fd9877871eca
Change-Id: I058c8f201fbff8d697ae030fb5d49867c21502d2