[roll] Roll fuchsia [gn_desc] Move file parsing into optimized library
This adds a function to the always-optimized gn_json rlib
so that all users can easily get the performance of a
release compilation of serde, without having to enable that
on their own binaries.

Tested by temporarily setting the binary back to not being
optimized, and verifying that there wasn't a performance
regression in a simple query to summarize a single label.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/825072
Original-Revision: 79ca513c1e7ef0d7447d198151474e9bafe75d6f
GitOrigin-RevId: 9624b9bf1dcc8bff8ef1b9216ea2fc68f3ac835c
Change-Id: Ida082b2b4cb2bde663be3e25bb7eed63a6763234