[roll] Roll fuchsia [a11y] Update responses only for held interactions.
This change ensures that a11y only tries to update a response of an
interaction if we previously put that interaction on hold.

Because of the nature of async code, it can be the case that an
interaction inside a11y becomes resolved before we send the initial
response to it, which caused an UpdateResponse call, trying to update a
value that was not committed.

Existing tests already cover this case, but can only be triggered in
specific conditions, so this change fixes potential flakes.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/825069
Original-Revision: 5adc33a1415fe213d79dd5cdbb1268cd14e34caa
GitOrigin-RevId: ec455ea7f629d8c8f2814aab67ad0febfcfa5a76
Change-Id: I69509fc8d14033972a364bd0405c59a5d701817f