[roll] Roll fuchsia [dfv2] Export to devfs through Node:AddChild
Add an API to export to devfs when you are adding a Node.

This provides the following benefits to DFv2:
- DF is now aware of which nodes correspond to which devfs paths.
- DFv2 drivers do not need to use the Compat API to pass around
  topological paths.
- Drivers can not lie about their topological paths or add nodes with
  incorrect paths.

This is a breaking change to DFv2 drivers that add a child node,
as their use of DevfsExporter might conflict with the entry added
by the child node.

RELNOTES: Please set DevfsArgs on NodeAddArgs to export to devfs.
Original-Bug: 117760
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791043
Original-Revision: 116a29acfb29753cb719e3e0d438817358c54e33
GitOrigin-RevId: 14bb47d86c787bb32b739ec8a3b85fb528f024e3
Change-Id: Ied6599deb674f259a59540b1d14e8dcca6583aeb