[roll] Roll fuchsia [fx][serve-updates] Fix checking for running ffx server
This updates the codepath where we are starting a `pm` server but are
configured to use the `ffx` server to stop checking the ffx config to
see the configured port. This is now unnecessary, since we can directly
ask the server which port we are supposed to be running on.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/825363
Original-Revision: 8c409700945cabad9087b8781ac95170d4af623f
GitOrigin-RevId: 0e451951a5932ac1acb017d6d43d534a78be05ce
Change-Id: I0771fd21e0074c8a9327dd4d9e2c4d13805cbda9