[roll] Roll fuchsia [arm64][crashlog] Try to dump memory around ELR when crashing.
Implement the feature request tracked in fxb/121735.  To summarize,
when we hit certain types of exceptions (esp. Unknown Exception), it
would be good to know (if possible) what the .text around the
exception link register address looked like.  This would help is to
either rule out or confirm that the exception might have been taken as
a result of a possibly corrupted instruction.

Every attempt is made to make sure that the memory should be readable
before we attempt to fetch, however as an added safety, there is a
chicken bit which is set and should suppress the feature in the case
that even attempting to read what should be valid memory results in a
double fault.

Original-Fixed: 121735
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/821178
Original-Revision: b9872b3356e0beaec82f637e71ab7e3f89efb27a
GitOrigin-RevId: af26661fc90876797f2c2a735b5a04d3c582bf06
Change-Id: I700029a4c9f34e4594d2877d90c62aead444ecd7