[roll] Roll fuchsia [fdf] Use sysmem FIDL in amlogic-video
Make the amlogic-video driver use the FIDL version of the sysmem
protocol instead of the banjo version. Amlogic-video depends on banjo
proxying in order to get the sysmem banjo protocol which DFv2 does not
allow.

Change the sysmen fragment in the vim3-video, sherlock-video,
nelson-aml-video, and astro-video composite bind rules to match to the
FIDL version of the sysmem protocol instead of the banjo version. The
amlogic-video driver binds to the composite devices that use those bind
rules and needs access to the FIDL version of the sysmem protocol.

Original-Bug: 123739
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/824271
Original-Revision: fc59cef1df92d05968a9aac6848ff5ea071b612f
GitOrigin-RevId: f92d8d73bc25395f9e7fc8d78bf236d3a94cbff2
Change-Id: I721ef2c3242a200e3360cb912b949e739d47c7cc