[roll] Roll fuchsia [starnix] Improve fidelity of Linux vDSO layout
This tweaks the vDSO linker script in two ways:

Updates the section layout to include the section headers in the
layout and allow the text (and other) sections to follow immediately.
This ensures that the section headers are in the memory layout and also
reduces the size to one page on x86_64.

Also includes the file header explicitly in the PT_LOAD section. This
works around a glibc bug where it assumes that the virtual address
offset of the PT_LOAD program header is always 0. Since this section
includes the file header at the start of the file the linker places
this section at a zero offset.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/825356
Original-Revision: 1d5ef5c002b098b15f8b144b90e36093cb042a45
GitOrigin-RevId: 3a29c99743ffbc19e8246db08a6ce345d9fabfcd
Change-Id: I11cab366434a837454c736513306827e92aaf203