[roll] Roll fuchsia [f2fs] Update 'od' invocation
The busybox od implementation for -i defaults to printing 2b decimal
values, so when dumping 32-bits with that vertion we get the string
'0      0' instead of just '0'.

Using '-td4' instead of '-i' will accomplish the same result but is
explicit in that output should be in 4 byte values.

Original-Bug: 123888
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/825074
Original-Revision: 309f45bccc79dfa9ddddf2e01cc35e49a6914d7e
GitOrigin-RevId: e803ec7a5b92c30716de829ee4633f857f3a1d36
Change-Id: I37a9acac9aafc75d8a88d8dcaad21ec226d18ab8