[roll] Roll fuchsia [ffx][emu] Using --length-is-lower-bound when resizing fvm
When launching the emulator, the product bundle metadata is read to
get the desired size of the fvm file. In some cases, such as building
the ASAN variant, the fvm is sized larger than specified in the
metadata. This would cause an error like

`Error resizing fvm: |length|(2147483648) must be greater or equal than
|disk_size|(3154280448 bytes)`

This change adds the -length_is_lower_bound flag to the fvm command
line so it will not be an error in this situation.

Long term, the meta data in the product should match the actual size
of the fvm.

Original-Bug: 102866
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/825222
Original-Revision: 787db5367f697b12709bfe037ce7908895857576
GitOrigin-RevId: 9e0132b5052ab6d428e83d4aea42442e3ad1abb4
Change-Id: I1ab6960b3558a9197fb241cf85542f212a3938cb