[roll] Roll fuchsia [devicetree] Introduce Matcher API
A Matcher provides a mechanism in which to perform "parallel",
walks of a device tree. That is, each matcher expresses interest
in a set of nodes/subtrees of the devicetree, and the infrastructure
notifies the matcher of a potential node of interest.

Additionally, a matcher may require multiple walks of the tree to
complete, in this case, the |Matcher::OnWalk| callback indicates the end
of a walk (E.g. all nodes have been visited).

A Matcher defines the maximum number of walks it might need to reach
completion, it is enforces that a matcher must complete before that
number of walks is exceeded. Failing to comply, results in a failure of
the |Match| operation, and the offending matcher |Matcher::OnError|
callback being called.

Original-Bug: 78675
Test: fx test devicetree-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/819247
Original-Revision: aae94d1a727bec8b388291e4be6153db76c24117
GitOrigin-RevId: f50e3b28d1302fae4f109d294efc1ff9f8d4af16
Change-Id: Ic99b5138405f02aaf5eacafd62e1edebc1ef7c86