[roll] Roll fuchsia [fatfs] Fix shutdown flake
I think this flake started showing up when we changed close so that it
just dropes the channel.

To fix this, we need to wait for the scope to finish.  I've also changed
it so that we have only one shutdown path now.

Original-Fixed: 123920
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/823867
Original-Revision: 66744bbb24ea4439f9b2232d48f262dbf9a12b6e
GitOrigin-RevId: d4d5024823b23c881ee8ce006b5aaa5f09400408
Change-Id: Ic197f393785d7f54109f43a2c5026ce02d9309b5