[roll] Roll fuchsia [ui][tests] Use a hermetic font provider for web_runner_pixel_test
Instead of a config-data package, this change configures the
web_runner_pixel_test to use a hermetic font server, which has
all the fonts it needs bundled in the package.

This allows us to stop using config-data, an approach that should be
obsolete in a CFv2 world.

Original-Bug: 123585
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/825593
Original-Revision: 4e24fdbdde485ec6416eda747300105173c17af1
GitOrigin-RevId: e5954656ead7145425292f32eac674108173d977
Change-Id: Ic09a86655eaf1c340328d4cc84cceb7b47424896