[roll] Roll fuchsia [ffx] Stop guessing if the current_exe is main ffx
The well-meaning attempt to figure out if `rerun_bin` was being run from
ffx or a subtool already had sharp edges, but since we distribute ffx as
a standalone binary called `ffx-<arch>` it causes very confusing errors
for certain users through no fault of their own.

Generally we know what kind of binary we're running from, though, so
just use that knowledge and add an extra field to `EnvironmentContext`
so it knows if the current binary is a 'real' ffx binary or not.

Original-Fixed: 124387
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/825818
Original-Revision: 45665b2439c211cb1611840ea15f7729d04ac396
GitOrigin-RevId: 3aa446f0c51045ee7527002491656a0c18b3f8ca
Change-Id: Ic37765d82475a2ca957606a0d64507992c345d1b