[roll] Roll fuchsia [i915][audio][hdmi] Allow TigerLake HDMI HDA to be seen, probed, and configured.
This contains changes to the Intel i915 HDMI audio driver
which allow the HDA audio to be recognized and configured as an audio
stream on TigerLake.  There is still no audio output, but the monitor ELD is recognized and the audio stream is published.  Note that the TigerLake support is disabled by default, and needs a one-line
developer change to enable the support.

Tested: Verified (on NUC11TNHv5) that audio-driver-ctl info returns the monitor's audio info if enabled, and that it returns the appropriate error if not enabled.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/824388
Original-Revision: 943bdc3912f4d1604d685f056bc80f762a558633
GitOrigin-RevId: cc8d9c4e54388bbde9f92bcc0ef382db4e27eb56
Change-Id: Ief861a3d57a8b8f4d192ac9cbc5c53f4ac63556a