[roll] Roll fuchsia [fidl] Experimental overlay feature
This is an experimental feature only available when the zx_c_types
experiment is enabled. Support is only in fidlc, not in any backends. It
is currently designed and implemented to support parts of the kernel
interface. In the future it may expand beyond that, but let's not get
ahead of ourselves, okay?

Overlays are a union-like structure. The syntax and JSON IR are very
similar to that of unions with a few constraints: no reserved members,
all overlays are strict, overlays may not be optional, no resource
overlays. The wire format is a 64-bit ordinal followed by the specified
member, padded to the size of the largest member.

IFTTT=no changes to good error tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/821868
Original-Revision: aa20df92be2960842f98c2e107eee259b9e1c9fd
GitOrigin-RevId: bcf5ea059df292ef16771b639899aec7a6e4e26b
Change-Id: I13d9a417d3feb80a15aa41d5ce33acb334545ef6