[roll] Roll fuchsia [starnix] Improve prctl(PR_SET_VMA, ..) support
This prctl can be used to assign a name to multiple mappings in one
call and has specific requirements on the input pointer and pointed to
string. The details of this prctl have changed in recent versions. This
updates Starnix to the behavior of a recent version (5.15).

This also propagates the name to snapshots (i.e. forked children).

Test: bionic_unit_tests_static, mm::memory_manager in starnix_tests.cm
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/799043
Original-Revision: 6db993e3f3333b4c782532aa4e3cbb6b8a280db6
GitOrigin-RevId: bb67120d44d078f2f92a780919fdcb1a5d76976c
Change-Id: I47f374927fa8a586e3ba158493c36ddf04048fc1