[roll] Roll fuchsia [perfetto-bridge] Fix Race in Startup
When starting up perfetto, perfetto-bridge uses a condition variable to
wait until perfetto has finished starting up. However, it's possible
that the Posted Task finishes and notifies the condition variable before
the main thread startings waiting on the cv, causing it to miss the
notify and wait forever.

This fixes the issue by adding an additional bool state protected by the
mutex. The state is only read or written while a lock on the mutex is
acquired. If the main thread sees that the state indicates that perfetto
is already initialized, it will instead skip waiting.

Original-Bug: 123637
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/825581
Original-Revision: 4832c51d566c778e52e2765fa453871192d1aae2
GitOrigin-RevId: 39f3bcd9cb10461a1ac291ac0b557693b549a9fb
Change-Id: Idaac779f04b6e8062fb8c3e8bb7d5e0f6e6031d8