[roll] Roll fuchsia [assembly] Move sysmem connector into assembly's core_realm AIB
This adds sysmem connector to all systems that have a '/core' realm.

The existing targets for adding sysmem_connector have been moved to
empty groups to enable the transition (and so that they can be removed
as deps in other files in a separate CL).

Original-Bug: 122862
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/825182
Original-Revision: 8698cfd60a000099e4060eb5474be17932dae21d
GitOrigin-RevId: 0fef9573b37f0347f1b84dc75afe4bf60644bb4d
Change-Id: I204d455c7f1b878e31e9bb3dfdddeca6aceaa6fe