[roll] Roll fuchsia [hwreg] Permit templated registers
There are many registers in the wild that are naturally modeled as being
templated. This change tweaks the internals of the hwreg DEF_* macros to
permit use within a templated register type. Further, the requirements
on that type are documented and made minimal: only `SelfType` must be
unambiguously defined within the applied scope.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/824543
Original-Revision: 56d8151a4a171ff269e8e0ef0303e2ade9424104
GitOrigin-RevId: 72b286df848c16206c6e1b49eb583dbe80a9057f
Change-Id: I43d70eb47525dc86bfff6c4a4f236b1b9a061cff