[roll] Roll fuchsia [starnix] Add a Close method on DevBinder
This will be needed when the protocol will be associated with a Task. At
this point, it will be necessary to correctly order open and close on a
given binder device, so that the same process doesn't end up opening the
device twice because the Open on the DevBinder has been reordered with
the close on the Binder channel.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/826021
Original-Revision: 659cbde780d0efb88ce7fb34b234247305fc3c99
GitOrigin-RevId: 87d8e4b660b728a1f5f607e6f2bce2e0c7e02a1a
Change-Id: Id684a3a82267f8344b590e715f8cbb9a91aa8fa1