[roll] Roll fuchsia [starnix] Fix underflow bug in sendmsg
Due to rounding up the read size to a multiple of sizeof(usize), it can
be the case that the size of "read" data can exceed the total amount of
data. In this case, the remaining size calculation would underflow and
usually lead to an allocation failure.

Original-Fixed: 124385
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/825079
Original-Revision: 5a481ce9bf883d3401c4fd3297d24b1845ded786
GitOrigin-RevId: ec7d4e52937d4cff42345662aecceea814b10caf
Change-Id: I781badb5eb8bcdcac5d94bb9bef0c4010d467779