[roll] Roll fuchsia [netstack/bench] Fix throughput calculation
The throughput calculation was erroneously subtracting 'lost_packets'
from 'packets', but the value associated with the 'packets' key in the
iperf3 JSON output means packets that were actually received, and not
total packets.

This incorrect assumption was made due to the fact that iperf3's
'lost_percent' value is calculated as `lost_packets / packets`,
so it would be natural to expect that 'packets' means total packets;
but in reality the 'lost_percent' value is not calculated as one
would normally expect.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/827079
Original-Revision: a34da145a1dfcc4f397855be26a11527c8876db7
GitOrigin-RevId: ad67bf26fb19bbc326faf0518ed81baae3797389
Change-Id: If1db67d65fbc73686df7e3e1602313d2a2f5593f