[roll] Roll fuchsia [audio] Fix virtualaudio/Control/GetNumDevices
Previously, virtualaudio/Control/GetNumDevices returned the count of
devices that had EVER been created; it did not reduce this number as
device instances closed. This is incorrect behavior. Now the FIDL
API returns the number of currently active (bound) virtualaudio
instances.

Because individual Device instances are unaware of the singleton
Control object, it is not notified when a Device closes. Instead, it
checks whether each device is bound (and cleans it out, if not) in
the GetNumDevices call itself (since this is the only time we
iterate the entire set of devices).

virtualaudio is only used by test code and developer utilities.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/827108
Original-Revision: 24a7baeb8a740781d3f6d49b48b513a48c5078f4
GitOrigin-RevId: 0d8af036e7fa3318af5edc3f335b3280d99cbcdd
Change-Id: I08b89b768ab4b784636d42ded966984cf14a5544