[roll] Roll fuchsia [fxfs][blob] Wire Fxfs into assembly
Adds a board `x64-fxfs` which instructs assembly to generate an Fxfs
image rather than an FVM image.  This Fxfs image contains a `blob`
volume which stores all of the blobs required for the system image.

Some changes were necessary to `fx qemu` to use this Fxfs image when
it is generated.

This change also permits Fxfs to use the Vmex resource which is
necessary to make the x64-fxfs board boot.

Original-Bug: 122056
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/806092
Original-Revision: 224cd72046e56300545d13e3bda5a685fb9c5379
GitOrigin-RevId: fce4b1b3b2003407f45ba9f3744f55bd52fd32be
Change-Id: I7872d38c9eee91d76308da55f98a85e7330656b5