[roll] Roll fuchsia [driver_runtime] Create thread pool and set role profile.
For now, we will create a new thread pool when the user
creates a dispatcher with a scheduler role. We will post
a task to set the role profile on that thread.

I've also updated the existing tests to not pass in erroneous
role profiles.

More work needs to be done to be able to elevate a dispatcher
to another role profile.

Original-Bug: 85946

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/827025
Original-Revision: 2e51ffbff0b42ffde02ab72fd52108947a45e7a6
GitOrigin-RevId: 069a151b7be318fb43a16308fbfbe048e3759aeb
Change-Id: I04c61c7a574b40e4cad0adeb0d4ef42621078a5c