[roll] Roll fuchsia [netstack][zxio] Set dgram write threshold to UDP max
Currently, clients sending data using Fast UDP can end up busy-looping
when using blocking sockets or waiting for the socket to become writable
with an API like poll. This occurs when the zircon socket is not
completely full (and is therefore writable) but does not have enough
capacity to fit the outgoing payload.

Change zxio's implementation of datagram_socket::wait_{begin,end} to
internally use ZX_SOCKET_WRITE_THRESHOLD instead of ZX_SOCKET_WRITABLE,
and in the netstack, set the socket's write threshold to the maximum
size of a UDP payload + the size of the TX metadata. This ensures that
when a client waits for a socket to become writable, the next write will
succeed as long as the payload size is within the maximum.

Original-Fixed: 123626
Test: fdio-zxtest, zxio-unittest-package
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/821907
Original-Revision: 1b32c793328e591a2d435fb1de1f59562fb374ef
GitOrigin-RevId: 5c08bd2fae7eb889afd48249ce249eb933ca706c
Change-Id: I6523c9c4b9a935feb60df2f81d3dc34202ff3b5c