[roll] Roll fuchsia [zxdump] Fix and test Thread::read_state
This wasn't tested and so the dump-via-TaskHolder changes
accidentally removed the actual support for dumping thread state
altogether because it had never been implemented in the live-task
API.  This implements it in the live-task API, and also refines
the Thread API to be more useful for examining the actual data
types.  It's tested via dump tests, which covers both the
live-task API indirectly by requiring it for dumps to have the
data and the dump-reader version of that API for verification.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/825359
Original-Revision: 5b2833fc282d86b3f582354c35469bccf5528811
GitOrigin-RevId: 46f30e010840d9b173bf56eab92fdc3dadb91cda
Change-Id: Ice77065e7c0f4dab9e52da6fbdfe863c87fa0e23