[roll] Roll fuchsia [starnix] Permit nanosleep with a relative REALTIME offset
The nanosleep syscall is supposed to specify a REALTIME offset and newer
libcs will implement the nanosleep() call with a clock_nanosleep() call
that explicitly specifies this clock. For now, Starnix assumes that the
REALTIME and MONOTONIC clocks advance at the same rate so we can
interpret a relative REALTIME sleep as a relative offset to the
MONOTONIC clock. This updates sys_nanosleep to use the correct clock and
relaxes the sys_clock_nanosleep implementation to permit relative
REALTIME offsets.

Original-Bug: 117507
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/827089
Original-Revision: 09eb3a8d9f0a632aa24039807f45018ad73adba0
GitOrigin-RevId: a4ee59975230d88752e21e4a26461d9a2c1ab930
Change-Id: Ie1810a95da2dafdf2a1da1159c5b137bb1e5314b