[roll] Roll fuchsia [kernel][phys] Convert elf-code-patching-test to physload module test
physload_module_test() gives a convenient framework for loading-related
phys tests: in this change we convert zbi-elf-code-patching-test over to
being specified via the template.

Also addressed in this change was a deficiency in which Symbolize
tracked not the current module, but the initial main one: this made it
difficult for intermediate nodes in a longer load chain to verify that
the expected parent nodes (via PT_INTERP build ID assertions).

Original-Bug: 91187
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/795906
Original-Revision: 55c3b2693356148b92d1dc103c1a4e9da4552d7b
GitOrigin-RevId: 74848df1a5dbaa7902f8ef7b0758c4ab1f8f2416
Change-Id: I6793285f3d2093d1f86aec8a0f57eeb4e8e3a550