[roll] Roll fuchsia [ffx][emu] Sticking the landing of Using instance data for target collection"
This is a reland of fxr/807069 and fxr/814450

The root cause of the issue was a race condition when starting and
stopping emulator instances without using isolate_dir. The target
instances were being matched by name, but the ssh port was different
for the different instances.

The symptom is the emulator starts in under 2 seconds (normal is ~20
seconds), and then ffx target list may show the instance, but
`ffx target show` or other commands that connect to the device
will fail.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/818547
Original-Revision: a58dfa820d85fd75c511f522594738aa6a3c95b5
GitOrigin-RevId: 9ee90923f0fd015771d5308cd940dceda940d98a
Change-Id: I1afe88f30da67c8ba7bdfe52a452821f1982435a