[roll] Roll fuchsia [netboot] Consistently namespace under "netboot"
This change is a pure rename: apart for some minor name tweaks for
clarity, this change ensures that all netboot definitions are prefixed
with either "netboot_" or "NETBOOT_". This is an established convention
for C libraries - and will be what we will want out of C bindings once
we FIDL-ify the protocol.

Original-Bug: 91102
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/821136
Original-Revision: 039c9ed8f501d896e4508e2ad2f06fe2962fab4c
GitOrigin-RevId: 8f6e12d6b150d0abeaff679be161463e3310fa12
Change-Id: Ieec54114c12ca6fccdf4420500ef3e23616e7160