[roll] Roll fuchsia [zither][rust] Annotate bindings with derived traits
These basic traits enable conveniences taken for granted, and thus are
prerequisites for generating useful bindings. The default traits derived
for everything are Copy, Clone, Debug, Eq, and PartialEq. Additionally,
zerocopy::{AsBytes, FromBytes} are derived where supported, which
streamline conversion to and from byte slices.

AsBytes and FromBytes require knowledge about the whole layout in order
to determine support. So that (a) this does not introduce logic to
reprocess each declaration's 'dependencies' in the generation of its
binding, and that (b) we do not complicate the main zither summarization
logic with backend-specific bookkeeping, we introduce a scheme so that
backends can provide a callback to be passed to zither.Summarize for
additional bookkeeping.

Original-Bug: 91102
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/822048
Original-Revision: a133ed2897d7a75d29e3acfae42d818c8fb51ed3
GitOrigin-RevId: 3f73949abd154b870a207b0cc743c94723aec61c
Change-Id: Iad9e395932a0f4daec371c20fa3910a2b18dbd24