[roll] Roll fuchsia [dfv2] Fix OOM test and put back in buildbot
Fixes the OOM test on DFv2. This test looked for a specific log message
that was dropped when the dfv2 main was separated out from v1. This log
is not necessary since there is another one that gets looked for the
comes after it. This fixes the log mismatch. Also puts back the test in
the buildbot now that it passes.

Original-Bug: 108569
Tests: fx test src/tests/oom

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/827881
Original-Revision: 44136c96d65659b7d477781134543047fe1c0764
GitOrigin-RevId: 73dbcae2e661bcdfdfc1af5d64f7b2ae29bfa325
Change-Id: I5b12d766b70bd533d603ba1d1fc5cd45f8d8df8d