[roll] Roll fuchsia [audio] virtual_audio_util: add usage and README
This CL adds to the virtual_audio CLI utility a --help flag that
details its various cmdline options, and a README that answers
common questions.

The utility now explicitly checks whether a specified initial
clock-rate is out-of-range, and fails in that case. (Previously
it did this only for rate-adjustments made once the device was
active.)

Small adjustments are also made to the available format and
gain options, and the default '--write-rb' value.

The CL also includes a number of 'lint'-type fixes from clang.

Test: diagnostic CLI utility for test fixture, no product change

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/827496
Original-Revision: be3cd7f433f77b443fc44bba45f5062fff1e9276
GitOrigin-RevId: 5d88dab7fdc78e48c647477e2be7542eac1f198b
Change-Id: I72c0fe78de039eb7e88408aa8c3e58f080c8c7ea