[roll] Roll fuchsia [starnix] Fix signal trampoline on aarch64
This changes the permissions and name of the signal trampoline. Both of
these were causing test failures, and since the mapping is only
temporary (until we get a vDSO), this change updates the permissions and
name instead of a more sophisticated fix.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/828496
Original-Revision: 924a4c040f5e1e9e74322b02413f226842083027
GitOrigin-RevId: e00054b1bf0ec004c226f3be418a88bbea7ab62d
Change-Id: I3e19e3b5bc73e0e0746f17133128698ff78e0586