[roll] Roll fuchsia [assembly] Add 'userdebug' AIB set for assembled_system to use
There are some artifacts that are needed on both eng and userdebug
builds and so they need to be included in non-user contexts, but they
aren't as restricted as the eng-only artifacts are.

This allows their use in userdebug systems.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/827936
Original-Revision: 2c2cd197309f1b4c9c3b5baa2f790d73ea8957f2
GitOrigin-RevId: fa1093e27569b32783ac16aadb74161a23b9f2db
Change-Id: Ib48582dfa9bd2d42326d21b14cb46a082dcad703