[roll] Roll fuchsia [ums] Simplify and re-eanble ums-integration-test.
This CL removes the Device protocol in
fuchsia.hardware.usb.peripheral.block. The protocol isn't actually used
to enable/disable the writeback cache in the ums-function driver (there
is no writeback cache). More importantly, the ums-function driver is
used for integration testing of the usb-mass-storage driver, and the
protocol is superfluous in that regard.

The test case UmsTest.CachedWriteWithNoFlushShouldBeDiscarded is flawed
in that it actually verifies that the supposedly cached write is not
discarded. This CL removes the test case.

Original-Bug: 121731
Test: fx test ums-integration-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/825071
Original-Revision: 8ccc0fbc6ca093ac0a5b52920f956c2630cf6a6b
GitOrigin-RevId: 939c242b374817a0219852b5c71d062e92c873e6
Change-Id: I09c88a1e923a4df05a3d2b8062a9896cdc2d9003