[roll] Roll fuchsia [fasync] Add shutdown call for emulated handles
Because emulated handles live in the process using them, not the Zircon
kernel, writes buffered in emulated handles go away when the process
dies.

This patch adds a call that ensures all emulated handles are flushed.
It at least guarantees whatever subsystem is on the other end of the
handles (i.e. Overnet) has seen whatever was in the buffers.

Original-Bug: 124614
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/828159
Original-Revision: 64951c2f7060119a29ec872ab1076b2513b335a8
GitOrigin-RevId: 71dc1901bccacdf2c7c93fc023fbb6eb0235ecd3
Change-Id: I74587e2813e10aa7d105063f0f84900a1bd391cc