[roll] Roll fuchsia [kernel][vdso] Fix build nits for GCC
Regressions in the build system long ago prevented the vDSO from
being built with GCC when the kernel is, as it originally was.
Subsequent changes will resolve that build system regression, but in
the interim some problems snuck into the vDSO code so that it is no
longer usable with the GCC tools. This change fixes one C++ bug
preventing compilation of the vDSO C++ code and another linker
script bug preventing runtime use of the vDSO in some places.

Original-Bug: 69697
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/829036
Original-Revision: c4cbbde1c8dd4677c19be2aa4c3435a49057d5cf
GitOrigin-RevId: b0e6af5343d2d0daeff4337fd3b6e0b2042dfb3c
Change-Id: I782fde151814f39c702d50c0a356591607c15e7e