[roll] Roll infra/recipes [firebase-docs] Build for target XOR host
If there is a version of a crate built for the target, don't build it
for the host in addition. Do this by grouping targets that are identical
except for prefix and only building for the first one in each group.

Original-Bug: 76722
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/828437
Original-Revision: 3de1b5bcc5479f05fe249820e4cb824ac606e89b
GitOrigin-RevId: edebc0dbc86cbb299fa2f693fbc7e66de9a1907c
Change-Id: I5523e84ffa269869a9e89833b14bd2df4c324d78