[roll] Roll fuchsia [brcmfmac] Don't glom tx frames in FW recovery
When the driver reloads the firmware in recovery, the firmware does not
expect frames it receives to be glommed, which changes the structure of
the frame. The driver later enables glomming in the firmware by setting
an iovar, but when initializing the firmware in the recovery worker,
frames should not be glommed.

Original-Bug: b/274966334

Test: Manually crash the firmware with the crash iovar, check that
firmware recovers and that user can connect to AP and ping.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/828137
Original-Revision: 9d30357b281c8cc2834e5bd018eb4c221e2d3f02
GitOrigin-RevId: 9ceb01b979a802f8a04bbe11fe4c7b3facdeddc9
Change-Id: I4a80beee27b83183ff55c389ea127f54d3a6cf33