[roll] Roll fuchsia [perfetto-bridge] Rig up Shared Memory in PerfettoProducer
This sets up perfetto producer to write perfetto trace events to the
shared memory buffer it is given as well as implements enough of the
perfetto::Producer API to handle the ipcs to start/stop tracing.

This means that:
- Perfetto producer no longer needs to be a trace observer
- We can write perfetto events to a buffer and funnel them to the trace
  controller

This doesn't yet connect to the perfetto tracing api yet. This directly
writes protobufs to the buffer rather than through the perfetto TRACE
macros.

Next steps are to:
- Figure out how to have the perfetto trace macros use the trace writer
  we create
- Abstract out the perfetto::Producer implementation to be usable
  outside of this test

Original-Bug: 111976

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/827879
Original-Revision: fac8d453630c9ee9360b6359c77ffa08807594d5
GitOrigin-RevId: 28ae6d8a7e1d261ba728210ae44f1d426dad5681
Change-Id: Ieb55e5ff52a5bdb61d66a23a7ffff78dc2fd88e6