[roll] Roll fuchsia [intel-i915] Descriptions for relevant Memory Controller registers.
The on-processor Memory Controller, also known as the MCH (Memory
Controller Hub), has its PCI BAR aperture mirrored as read-only in the
graphics BAR space. This is intended to make it easy for the display
driver to learn low-level DRAM configuration details needed for
memory bandwidth computations, and for configuring the arbiter.

This CL adds definitions for the MCH registers that are relevant to
display engine configuration.

Original-Bug: 111612
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/792122
Original-Revision: ba60f29459a67ce9a9935a4a769a4b89a7696d22
GitOrigin-RevId: 6a9f06288c201178ad5865e5920d76387f4bdcfa
Change-Id: I3b73b863b00f901e8cc8f00749d32c0cae59177b