[roll] Roll fuchsia [scenic][tests] Fix flaky assertion in tests on session teardown.
When a Flatland Session is torn down, Flatland manager removes the
session ID from list of sessions, and posts a task on the session's
loop thread to delete Flatland instance, which triggers Flatland
Presenter client calling Presenter.RemoveSession() FIDL method,
while the presenter may be handled on another loop.

Our tests asserts that a session is torn down once the session ID
is removed from the Flatland manager, and then expects the
RemoveSession() FIDL call to be called. As mentioned above, this
might not be true because of the race between threads.

This change modifies the waiting condition of FlatlandManager
main loop in tests (so that it waits for the Flatland instance being
destroyed and the RemoveSession call is handled). This doesn't
affect the behavior on production code. Besides, we also add a
fit::thread_checker to the list of removed session IDs, to make
sure it is always invoked on the test loop thread for thread safety.

Test: flatland_unittests on core.qemu-arm64 (without KVM)
Original-Fixed: 124632
Multiply: flatland_unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/828577
Original-Revision: 1a8692f18d1eecdf13e7d960132238e0fcbc31e0
GitOrigin-RevId: c7266c6287952241a8ab5818e37ae748cca3c722
Change-Id: Ie61b58ecc2fdc9d7d3c97de64cb2f474cfdc87b6