[roll] Roll fuchsia [radar][proxy] Add connect request queue to fix startup flake
If a connect request is received immediately after radar-proxy starts,
there is a chance that it will be handled before the existing radar
devices are enumerated and a connection is established. This will cause
the connect request to be completed with an error even though radar
devices are available.

To fix this, a queue is added to hold connect requests that arrive
when there is no connection to a device. Once a connection is made, any
requests in the queue are completed using the new device.

Also re-enable radar-integration-test now that the issue is fixed.

Original-Fixed: 124621
Multiply: radar-proxy-test: 1000
Test: radar-integration-test passed on Nelson tryjobs
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/827220
Original-Revision: cecfa4a69abded19a993df2b9c40c1dbe3d8974b
GitOrigin-RevId: caccbffefcae2bdd8c16fd5d63525067582db196
Change-Id: Iff047eeebfac13d6c3c68457565ec2a38489d849