[roll] Roll fuchsia [gigaboot] Add an indication that boot mode skips normal boot
When trying to boot Fuchsia on a NUC I've ran into an occasion several
times where the NUC refuses to boot and goes straight into fastboot. I
was only able to recover this by entering the UEFI shell and running
`dmpstore -d bootbyte`, to delete the boot mode variable.

I'm not familiar with the mechanics of this variable. This CL adds a
note during the boot process when that is the case.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/829642
Original-Revision: a8bd213aa8c84d374f870f92c8a8053d5b128d7c
GitOrigin-RevId: 2770d6f5901738314a4743fec568d5a9b850501c
Change-Id: Ib3a95265ca44fc80cd021dec48c05696460e9d0e