[roll] Roll fuchsia [reachability] Improve clippy linting for async
This is a minor bug, and according to
rust-lang.github.io/rust-clippy/master/index.html#unused_async it is
beneficial to check for this and remove it to reduce mental and
computational overhead with this declaration.

This CL adds a warn clippy tag for functions that hae an async tag
without having any awaits within the function body.

With this lint rule, two cases were caught and fixed:
- `resume` in NetworkChecker
- `handle_ten_secondly_telemetry` in Telemetry

Fix: 124592
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/826659
Original-Revision: 4c423b7965aec86841410bfd2964fd21ff4cebdd
GitOrigin-RevId: f7529d0c4402fe1578f31de5e70194f2f7e5adb9
Change-Id: Id9d52ddd8a2509a0a2e3226a363abb3d000a80d4