[roll] Roll fuchsia [netstack3] Send ethernet frames over loopback
Support Linux-style packet sockets by treating loopback as an ethernet
device that can send frames with source and destination MAC addresses.
This is invisible to the rest of the netstack but will be exposed via
packet sockets. As part of this change, rename queue-related things to
be frame-oriented or ambivalent.

Original-Bug: 106735
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/828237
Original-Revision: b6298debe3c8b86c045b324de81e8631e6825ff8
GitOrigin-RevId: bf723db50e700f24e8886d271e3da524762bfe41
Change-Id: Ifde2c2c2600d4aaf67f841937f98ad82b4c538bc