[roll] Roll fuchsia [fxfs] Simplify checking of directory children
Prior to this change, fsck followed different paths depending on the
order in which objects were encountered.  This change makes it so that
directory children are processed in a second pass (the same pass that we
currently scan extents in), and allows for a few simplifications.

This CL also includes some other changes:

 * Directory::update_attributes now preserves the operation which
   is required to ensure object counts are correctly updated.  It has
   also been simplified.
 * Directory::insert_child will now update the sub_dirs field when
   inserting a directory.
 * The algorithm for checking link cycles has been changed so that it
   now catches the case where two directories refer to each other as
   children.  This change in algorithm means we no longer need to record
   all the children of directories (which should reduce memory usage).

Original-Fixed: 124705
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/829136
Original-Revision: 02d6273e3a214c19168087a278d5afa62586146f
GitOrigin-RevId: a2587b81187173254922c2654b83d0a95e9f46d4
Change-Id: I81bcbfc1ed24bb5d20d0488bed5ffd7f311c184e