[roll] Roll fuchsia [netstack3] Implement timeout for defunct FIN_WAIT_2 connections
FIN_WAIT_2 is a potential state that can stuck forever in the system:
if the peer end just dies, and the connection becomes defunct/orphaned,
there is no way that we will ever look into the connection; in other
'closing' states, we either are waiting for a `close` from the user or
we are waiting for an ACK, whose absence will timeout the connection.
To solve this problem of lingering FIN_WAIT_2 states, we need a timer
for the state that is not described in the RFC, but is de facto present
in most implementations.

Original-Bug: 117956
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/827456
Original-Revision: 9c345ec9185e5f3e642b4238cef1cf6816662c02
GitOrigin-RevId: deda48ecaeac81c875c8a8ba2f316a01e1694815
Change-Id: Id38b1ce3220aefcc0c7442a775bee1891637a9c9