[roll] Roll fuchsia [starnix] Implement the remote binder driver device.
The device expects a userspace process to connect to it and use an ioctl
to transmit the name of the service to use to start a remote fuchsia
component implementing using binder.

The driver will connect to the remote process using the RemoteController
protocol, and will drive the local process so that local task are
associated with remote threads when doing binder ioctls.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/823787
Original-Revision: dfbcb1d6aba097ef20fa01ec29c0fd89df4066a6
GitOrigin-RevId: aaf916e223296b4afae17009c29b51103dc2e899
Change-Id: I5e65fb04c19b35aec9d1ddda7942c919684d1c45