[roll] Roll fuchsia [build] Add clean build fence to work-around Ninja incremental failures.
The change in [1] was submitted successfully, but it triggers
flaky incremental failures, due to the fact that it changes the
build graph in a way that Ninja cannot deal with properly
(see associated bug for underlying reason).

This CL introduces a clean build fence to force clean builds
after it is applied, to get rid of the issue completely.

Original-Bug: 83397
Original-Fixed: 124905

[1] https://fuchsia-review.googlesource.com/c/fuchsia/+/829176/1

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/830162
Original-Revision: 32bcabf65678e30d5c6c9ff2957c1f361e8e06e8
GitOrigin-RevId: 8f051738480622a57ace1ef5f89ffe8e3bd0447f
Change-Id: I6f9c097708549665ac1d7e054194980345a04b96