[roll] Roll fuchsia [kernel][no-userabi] in the kernel_no_userabi case turn off some code
A few places that assume some symbols are declared in lib/userabi are
referenced in VM and unittest code. Add a #define to the kernel and
switch off some code based on it.

Needed to link a functional kernel with the switch set.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/822200
Original-Revision: 8b380c92144c2bf1c47266750bf4dad7937b5ebb
GitOrigin-RevId: 54832baf410cb0999e790e24724aaff0c3c2beb0
Change-Id: I295494bd0edd90d5be78c1faaced60a86dc486b6