[roll] Roll fuchsia [bt][hfp] Connect CallManager on reconnecting peer
When reconnecting a peer, the peer task would be started up, but this
would leave the peer in a state without being aware of the currently
active CallManager, causing calls to never be transmitted to the peer.

This most often was seen as a failure to use HFP, and instead fallback
to A2DP and the internal microphone.

Original-Bug: 96493
Original-Bug: 105082
Test: added unit test for PeerImpl, fx test bt-hfp-audio-gateway-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/822235
Original-Revision: 49260199cdee19cecb4f40135bb302fdc5a5af55
GitOrigin-RevId: 42bd2211acf10b1d168e5e45fe8b69c0eb79a88e
Change-Id: I53b2196df4f5e3e734491e614367df405b2b8138