[roll] Roll fuchsia [mdns] answer instance-related questions using correct sections
When answering questions, InstanceResponder was always putting the
PTR resource in the answer section and everything else in the
additional section. This is correct for PTR queries, but not
for other queries.

This CL updates the InstanceResponder so that it puts the resource
matching the question in the answer section and other resources
in the additional section. When responding to ANY questions, the
PTR resource is the answer if the service type is specified, and
the SRV and TXT resources are the answer if the service instance
is specified.

Original-Fixed: 104937

Test: new unit tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/813967
Original-Revision: 3855079ecb3ae7331a8984a0188b1dd6407ba7c9
GitOrigin-RevId: be108aaf22bea6762ae88d0bb119438ff1e2c5ba
Change-Id: I1073a7f7366819be4b28310b2061512c87e8581c