[roll] Roll fuchsia [codecs][sbc] Send queued audio when space is low
When space in the output buffer is low (less than enough to store
another packet output), send the queued output, instead of relying on
the output matching exactly the correct size.

Not doing this means the SBC decoder sends an error status because there
is not enough space in the output buffer (and drops the input frame).

Original-Bug: 122271
Test: updated tests to add some buffer constraints
Test: fx test audio_decoder_test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/816930
Original-Revision: 448908146c0245523cc2dbb0859e9ecdda3e49c8
GitOrigin-RevId: 1548ace0caeae1b6f2152e47b5ed64e7a8318304
Change-Id: Ic60c3cc62ae26781a967d1c2ac58f8d9ea4d23c6