[roll] Roll fuchsia [sdk] Remove obsolete modular SDK GN templates.
The Modular SDK is obsolete, so this CL removes the last
templates related to it from the Fuchsia build.

Note that this also removes //sdk/tools/BUILD.gn. This file
seems to be completely unused today, and no other target in
the build graph depends on anything it defines.

Original-Fixed: 114745
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/829259
Original-Revision: c31692fcac38dd14b370f78188cc6f0ce606230a
GitOrigin-RevId: 6295f977c64496e7c996fa4d07b0e9eb9ec89601
Change-Id: I3821220f5c5d3ea7e6918f00fbab49377919eb54