[roll] Roll fuchsia [kernel][riscv][user-copy] fix user copy routine
Previously this routine was trashing s2 register, which is now used for
the shadow call stack. However it was also relying on the memcpy routine
being compiled a particular way, which it no longer is, due to frame
pointers and shadow call stack.

Though it's not a complete fix, tighten up the code that wraps memcpy to
try to protect against most registers being trashed. The better solution
is to have a fully asm implemented memcpy/memset routine so that up
front we known what registers are touched.

Also clear the sstatus.sum bit on kernel entry and add some additional
validity checks on exception entry to try to catch things getting out of
place.

Do not attempt to process the page fault if it originated from kernel
space and isn't part of a user copy.

Original-Bug: 122873
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/823109
Original-Revision: 92790bc675c1d20da6c851e81c78621e8f73ed9e
GitOrigin-RevId: a85de2de95ce3aa7c9a388369826a8165fc6271d
Change-Id: Ie4029d979e70e16147912ae75951e01c2c8b3faa