[roll] Roll fuchsia [fdf] DriverRunner removes orphaned node after binding
Update DriverRunner so it stores orphaned nodes in a
map. When the DriverRunner binds a node, it will remove
it from the orphaned nodes.

This is to address an edge case in which a node fails
to match to a driver, then binds to one after a
RequestBind() call.

Original-Bug: 123392
Tests: Added a test to driver_runner_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/828384
Original-Revision: 090004cc69acac34bf9d3b02b6debeea58bc7d2a
GitOrigin-RevId: 708bf963249b8aedf2f92d0e55d2b360ccf1c2f1
Change-Id: I55f6ab1d647d4e46af226b1044f1679ebd05fa07