[roll] Roll fuchsia [kernel][riscv][brwlock] disable the PI version of the brwlock for riscv
At the moment, the PI version of the brwlock requires the use of a
doubleword compare-and-swap operation which does not exist on the RISC-V
architecture. For the moment, simply disable using the PI version of the
lock for this architecture by defining BrwLockPi to actually be
BrwLockNoPi.

Original-Bug: 122873
Original-Bug: 124881
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/824298
Original-Revision: 34211560c1048becc521329766ca19893c2529cf
GitOrigin-RevId: b3550efbce60e1b23caf70a51af4cf8dde84e9d1
Change-Id: If93d08c00a7ca46ab965c518a8006473a48cd5e3