[roll] Roll fuchsia [testrunner] Reduce SSH reconnect timeout.
Before, we automatically retried to connect if we lost SSH connection
during a test with a timeout of 2 minutes. This would allow us to
reconnect even if the target underwent a reboot. We should reduce the
timeout so that we can report failures due to the target rebooting so
that they can be investigated in case the reboot is caused by a certain
test.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/830664
Original-Revision: a1d13a944dacf2ac279498c3998852811a2b5cb7
GitOrigin-RevId: 52293fab4a1c7cda0c3013e9ddeab362c5ab357e
Change-Id: I9a7cb76baed246377323de850d89811d6b276391