[roll] Roll fuchsia [diagnostics][triage] Delete `Multiple` error
Triage-lib has a kind of error called "Multiple" that's a collection of
other errors, so every error can be seen and acted on.

In practice, this causes code clutter and also clutter in output. We
should treat an error more like a thrown exception (short-circuiting
evaluation), not assume that every error needs to be user-facing.

This CL deletes the Multiple error type entirely, dropping any errors
beyond the first.

It does not delete the "Ignore" error list - that may be a good idea,
but it's a separate issue.

Original-Fixed: 106637
Original-Bug: 58922
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/830778
Original-Revision: 23f65508ad2c8040aee6f3f6e5ab194beea94c6c
GitOrigin-RevId: 9a91babf96e46f16afe49995cfe830866edc0369
Change-Id: I0ca794a6e985b3e742fe68c6c0274ef40ef9680c