[roll] Roll fuchsia [ffx][target] Loosening network changed detection
The effect of returning true for network_changed from
find_matching_target is to trigger reconnecting the host_pipe
to the target. The main reason to do this is for user mode
networking with an emulator, where the port mapped to SSH changes,
but the host pipe is using the old port mapped to an old instance.

A side-effect of this physical targets that respond to
mDNS on IPv4 and IPv6, the address will change quickly
as the target is coming up, this can cause a lot of
confusion and race conditions.

This is the probable cause of failures in the Chromium CI/CQ tests
which use physical devices.

To avoid that, only return network changed if the port
is specified as well as a change.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/830316
Original-Revision: 83853cfc918caea4cdf5a4055e7fa6167f1f18c0
GitOrigin-RevId: 35349cbca55169c74c70e6847b22afda02e19d8c
Change-Id: I1c835140a4aa5ed14408d41492bb5b0c1e6e5b1b