[roll] Roll fuchsia [scripts] Run incremental publisher with time versioning
The ffx repository server has a TUF client, which tracks the versions of
the TUF metadata. We generally generate metadata where we use the latest
time as the version number, rather than monotonically increasing the
number. This allows us to do a `fx clean-build`, and the TUF metadata
to be correct.

Unfortunately we were running the incremental publisher without time
versioning. So if a user does `fx clean && fx test ...`, this can cause
a new TUF repository to be created, where the metadata version is 1.
TUF considers this a rollback attack, and will reject the metadata.

This avoids that case by enabling time versioning in all the places
where we use the incremental publisher.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/831298
Original-Revision: 7eded24645140e705401cb92b749a751b9e3d41a
GitOrigin-RevId: 9a5bc342f6d755d4c95bfb8367e00298a8c61fe9
Change-Id: I3c31e2909303405614502b48d22e7405669f488a