[roll] Roll fuchsia [a11y] Wires a11y manager with new gd protocol.
This change is part of the accessibility migration to use the new gesture
disambiguation protocol.

this change instantiates a touch source to be used by the a11y manager, and
passes it to the gesture manager.

the Screen Reader and magnifier then register their gestures to process the
arriving pointer events.

Test: fx test a11y-lib-tests
Original-Bug: 100030
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/801599
Original-Revision: fb126d6ec75cc1156f7048f26cbc64080eaa8b91
GitOrigin-RevId: 383930c925532f0f3fd4f5991f9c66df7b721fcc
Change-Id: Ie63fe9359d8be643150fa03662115275331fea85