[roll] Roll fuchsia [display][core] Enforce thread affinity of Fences/FenceReference creation and destruction.
Fence refcount implementation is not thread-safe (though thread
compatible). In today's display::Client implementation, all the
Fences and the References to Fences are created on the same
controller loop thread.

For the sake of simiplicity and to avoid data race, we add a
thread enforcement to Fence / FenceReference that they must be
created and destroyed on the same thread.

This change adds a fit::thread_checker which issues debug asserts
to make sure that methods of Fence and FenceReference always run
the same thread where the Fence is created. There's no behavior
change introduced by this test.

Test: display-core-unittest on core.x64-asan
Test: vkcube-on-fb on core.x64-asan
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/823687
Original-Revision: ffa8b74ad407ce08aec110f4f944382183aa1b2e
GitOrigin-RevId: bc0a2e6f37dd68ac908df50abc0386f62387a4f7
Change-Id: I83baa462974131c8b3c93849bcccc4d383822ce6