[roll] Roll fuchsia [zircon] Remove SYSROOT_ prefix from public header guards
Not only is "sysroot" a misnomer, but it features neither into the
include nor source paths of these headers, making it difficult to
generate these in a general way (as we intend to from FIDL).

This change should not open the platform up to header guard collision,
as absolutely nothing else in the platform should have header guards
matching that of the zircon public headers.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/828897
Original-Revision: a41c2de53392f9b97785d9650561b2ab733f7ee0
GitOrigin-RevId: adbd57e2a94dae0071fedb11fd242c07f46b6ab5
Change-Id: Ie4247ce6602bc9568bee42b713e45f157fc1b89d